### 2020-04-29 #80 @whedon
[PRE REVIEW]: Take a moderndive into introductory linear regression with R
**Submitting author:** @rudeboybert (<a href="http://orcid.org/0000-0001-7824-306X">Albert Y. Kim</a>)
**Repository:** <a href="https://github.com/moderndive/moderndive" target ="_blank">https://github.com/moderndive/moderndive</a>
**Version:** v0.5.1

@rudeboybert if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
