### 2020-05-10 169826c
styler::style_pkg(filetype = c("R", "Rmd"))
