+++ R/pennies.R
@@ -1,7 +1,7 @@
 #' A sample of 50 pennies
 #'
 #' A sample of 50 pennies contained in a 50 cent roll from Florence Bank on
-#' Friday February 1, 2019 in downtown Northampton, Massachusetts, USA 
+#' Friday February 1, 2019 in downtown Northampton, Massachusetts, USA
 #' \url{https://goo.gl/maps/AF88fpvVfm12}.
 #'
 #' @format A data frame of 50 rows representing 50 sampled pennies and 2 variables
@@ -9,41 +9,41 @@
 #'   \item{ID}{Variable used to uniquely identify each penny.}
 #'   \item{year}{Year of minting.}
 #' }
-#' @note The original \code{pennies_sample} has been renamed \code{\link{orig_pennies_sample}} 
+#' @note The original \code{pennies_sample} has been renamed \code{\link{orig_pennies_sample}}
 #' as of \code{moderndive} v0.3.0.
 #' @examples
 #' library(ggplot2)
-#' 
+#'
 #' ggplot(pennies_sample, aes(x = year)) +
-#'  geom_histogram(binwidth = 5, boundary = 2000)
+#'   geom_histogram(binwidth = 5, boundary = 2000)
 "pennies_sample"
 
 
 
 #' Bootstrap resamples of a sample of 50 pennies
 #'
-#' 35 bootstrap resamples with replacement of sample of 50 pennies contained in 
-#' a 50 cent roll from Florence Bank on Friday February 1, 2019 in downtown Northampton, 
+#' 35 bootstrap resamples with replacement of sample of 50 pennies contained in
+#' a 50 cent roll from Florence Bank on Friday February 1, 2019 in downtown Northampton,
 #' Massachusetts, USA \url{https://goo.gl/maps/AF88fpvVfm12}. The original sample
 #' of 50 pennies is available in \code{\link{pennies_sample}} .
 #'
-#' @format A data frame of 1750 rows representing 35 students' bootstrap 
+#' @format A data frame of 1750 rows representing 35 students' bootstrap
 #' resamples of size 50 and 3 variables
 #' \describe{
 #'   \item{replicate}{ID variable of replicate/resample number.}
 #'   \item{name}{Name of student}
 #'   \item{year}{Year on resampled penny}
 #' }
-#' @seealso \code{\link{pennies_sample}} 
+#' @seealso \code{\link{pennies_sample}}
 #' @examples
 #' library(ggplot2)
 #' library(dplyr)
-#' bootstrap_sample_means <- pennies_resamples %>% 
-#'   group_by(name) %>% 
+#' bootstrap_sample_means <- pennies_resamples %>%
+#'   group_by(name) %>%
 #'   summarize(sample_mean = mean(year))
-#' 
+#'
 #' ggplot(bootstrap_sample_means, aes(x = sample_mean)) +
-#'  geom_histogram(binwidth = 2.5) +
+#'   geom_histogram(binwidth = 2.5) +
 #'   labs(x = "sample mean year", title = "Bootstrap distribution of sample mean year")
 "pennies_resamples"
 
@@ -51,7 +51,7 @@
 
 #' A population of 800 pennies sampled in 2011
 #'
-#' A dataset of 800 pennies to be treated as a sampling population. Data on 
+#' A dataset of 800 pennies to be treated as a sampling population. Data on
 #' these pennies were recorded in 2011.
 #'
 #' @format A data frame of 800 rows representing different pennies and 2 variables
@@ -63,22 +63,24 @@
 #' @examples
 #' library(dplyr)
 #' library(ggplot2)
-#' 
+#'
 #' # Take 25 different samples of size n = 50 pennies from population
[…]
