### 2020-05-12 f9ebd67
Implement the fullrange option for geom_parallel_smooth
Allows geom_parallel_smooth to draw regression lines over the entire support of the x-axis by setting fullrange=TRUE, just like you can with geom_smooth.
