### 2020-05-17 1d21668
Update `geom_parallel_slopes()`.
- Include `level` argument.
- Refactor for more resemblance to {ggplot2} implementation of `GeomSmooth`.
