### 2020-06-11 907b854
basic categorical regresssion version of geom_smooth.
