+++ R/geom_categorical_model.R
@@ -102,7 +102,15 @@ compute_group_new_categorical_data <- function(group_df, scales, n) {
   x_seq <- seq(support[1], support[2], length.out = n)
   group_seq <- rep(group_df$group[1], n)
   
-  data.frame(x = x_seq, group = group_df$group[1], x_orig = group_df$x[1])
+  if (as.numeric(group_df$x[1]) == 1) {
+    linetype = 1 
+  } else {
+    linetype = 2
+  }
+  
+  data.frame(x = x_seq, group = group_df$group[1],
+             x_orig = group_df$x[1], linetype=linetype
+             )
 }
 
 predict_categorical_df <- function(model, data, se, level) {
