### 2020-06-11 9bf2bc7
added basic test cases for geom_categorical_model
