### 2020-06-11 806e574
don't try and set linetypes if there is already a linetype variable present
