### 2020-06-11 5e6fd4c
warn and draw nothing if given a numeric x axis
