+++ .travis.yml
@@ -28,6 +28,7 @@ notifications:
 
 r_packages:
   - devtools
+  - rticles
   
 r_github_packages:
   - hadley/pkgdown
