### 2020-06-12 b063870
Cleaned up documentation for geom_categorical_model
