+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export("%>%")
+export(geom_categorical_model)
 export(geom_parallel_slopes)
 export(get_correlation)
 export(get_regression_points)
