### 2020-06-17 60ce183
Update NEWS.md
