+++ NEWS.md
@@ -7,7 +7,7 @@
 * Update `geom_parallel_slopes()` with new arguments:
     + Use `fullrange=TRUE` to draw regression lines over the entire support of the x-axis (by @wjhopper)
     + Use `level` to set different level of confidence interval shading (by @echasnovski)
-* Added new function `geom_categorical_model` for visualizing regression models with one categorical variable (by @wjhopper)
+* Added new function `geom_categorical_model()` for visualizing regression models with one categorical explanatory/predictor variable (by @wjhopper)
 
 ***
 
@@ -89,4 +89,3 @@ Fixed broken url in `?bowl_samples`
     + `pennies`: 800 pennies to be treated as a population from which to simulate sampling a numerical variable from (`year` of minting)
     + `bowl`: Bowl of 2400 balls of which 900 are red to be treated as a population from which to simulate sampling a categorical variable from (`color`). Also known as the urn sampling framework \url{https://en.wikipedia.org/wiki/Urn_problem}. 
     + `bowl_samples`: data from tactile version of sampling from `bowl` done in class: 10 groups sampled n=50 balls from  and counted the number red [ADD MODERNDIVE LINK]
-
