### 2020-06-17 5c22f74
Update R/geom_categorical_model.R
