### 2020-06-17 5bbf1e2
Update R/geom_categorical_model.R
