### 2020-06-17 1b5f0fa
Tweaked documentation
