+++ vignettes/why-moderndive.Rmd
@@ -34,9 +34,10 @@ affiliations:
    index: 3
 bibliography: paper.bib
 output:
-  rticles::joss_article:
-    keep_md: yes
-    number_sections: yes
+# Uncomment the three lines below to build JOSE article using rticles:
+#  rticles::joss_article:
+#    keep_md: yes
+#    number_sections: yes
   rmarkdown::html_vignette:
     df_print: default
     number_sections: yes
