### 2020-06-17 049a07f
Rebuilt all help files with roxygen v7.1.0
