+++ DESCRIPTION
@@ -35,7 +35,7 @@ Imports:
     infer,
     rlang (>= 0.2.0),
     glue
-RoxygenNote: 7.0.2
+RoxygenNote: 7.1.0
 Suggests: testthat,
     covr,
     rmarkdown,
