+++ pkgdown.yml
@@ -1,7 +1,6 @@
 pandoc: '2.2'
-pkgdown: 1.4.1.9000
-pkgdown_sha: d5e60a7232ee5f414f2a1de2aca5bf44054e7a08
-articles:
-  why-moderndive: why-moderndive.html
-last_built: 2020-03-19T14:27Z
+pkgdown: 1.5.1.9000
+pkgdown_sha: caf7cc74e008e8d0d6e53eeda93cd3cad3545bf2
+articles: []
+last_built: 2020-06-17T18:43Z
 
