### 2020-06-17 1958ae7
update _travis.yml. fixes #80
