+++ pkgdown.yml
@@ -1,6 +1,6 @@
-pandoc: '2.2'
-pkgdown: 1.5.1.9000
-pkgdown_sha: caf7cc74e008e8d0d6e53eeda93cd3cad3545bf2
-articles: []
-last_built: 2020-06-17T18:48Z
+pandoc: 2.3.1
+pkgdown: 1.4.1
+pkgdown_sha: ~
+articles:
+  why-moderndive: why-moderndive.html
 
