+++ man/mythbusters_yawn.Rd
@@ -8,9 +8,9 @@
 A data frame of 50 rows representing each of the 50 participants
 in the study.
 \describe{
-  \item{subj}{integer value corresponding to identifier variable of 
+  \item{subj}{integer value corresponding to identifier variable of
   subject ID}
-  \item{group}{string of either \code{"seed"}, participant was shown a 
+  \item{group}{string of either \code{"seed"}, participant was shown a
   yawner, or \code{"control"}, participant was not shown a yawner}
   \item{yawn}{string of either \code{"yes"}, the participant yawned, or
   \code{"no"}, the participant did not yawn}
@@ -31,7 +31,9 @@ library(ggplot2)
 # Plot both variables as a stacked proportional bar chart
 ggplot(mythbusters_yawn, aes(x = group, fill = yawn)) +
   geom_bar(position = "fill") +
-  labs(x = "", y = "Proportion", 
-  title = "Proportion of yawn and not yawn for each group")
+  labs(
+    x = "", y = "Proportion",
+    title = "Proportion of yawn and not yawn for each group"
+  )
 }
 \keyword{datasets}

[…]
