### 2020-06-17 bde3346
Rebuilt README after updating R + pkgs
