+++ R/geom_categorical_model.R
@@ -1,7 +1,7 @@
 #' Regression model with one categorical explanatory/predictor variable
-#' 
+#'
 #' \code{geom_categorical_model()} fits a regression model using the categorical
-#' x axis as the explanatory variable, and visualizes the model's fitted values 
+#' x axis as the explanatory variable, and visualizes the model's fitted values
 #' as piecewise horizontal line segments. Confidence interval bands can be
 #' included in the visualization of the model. Like \code{\link{geom_parallel_slopes}},
 #' this function has the same nature as \code{geom_smooth()} from
@@ -15,31 +15,29 @@
 #'
 #' @inheritParams ggplot2::layer
 #' @inheritParams ggplot2::geom_smooth
-#' 
+#'
 #' @seealso \code{\link{geom_parallel_slopes}}
 #' @export
 #'
 #' @examples
 #' library(dplyr)
 #' library(ggplot2)
-#' 
-#' p <- ggplot(mpg, aes(x=drv, y=hwy)) +
+#'
+#' p <- ggplot(mpg, aes(x = drv, y = hwy)) +
 #'   geom_point() +
 #'   geom_categorical_model()
 #' p
-#' 
+#'
 #' # You can use different colors for each categorical level
-#' p %+% aes(color=drv)
-#' 
+#' p %+% aes(color = drv)
+#'
 #' # But mapping the color aesthetic doesn't change the model that is fit
-#' p %+% aes(color=class)
-#'   
+#' p %+% aes(color = class)
 geom_categorical_model <- function(mapping = NULL, data = NULL,
                                    position = "identity", ...,
                                    se = TRUE, level = 0.95,
                                    na.rm = FALSE, show.legend = NA,
-                                   inherit.aes = TRUE
-                                   ) {
+                                   inherit.aes = TRUE) {
   # Possibly warn about not needing to pass `method` argument
   dots <- list(...)
   if ("method" %in% names(dots)) {
@@ -48,102 +46,100 @@ geom_categorical_model <- function(mapping = NULL, data = NULL,
       '("lm" is used).',
       call. = FALSE
     )
-    
+
     dots <- dots[setdiff(names(dots), "method")]
   }
-  
+
   # Construct layer params
   stat_params <- list(na.rm = na.rm, se = se, level = level)
-  
+
   params <- c(stat_params, dots)
-  
-  ggplot2::layer(geom = ggplot2::GeomSmooth, stat = StatCategoricalModel,
-                 data = data, mapping = mapping, position = position,
-                 params = params, inherit.aes = inherit.aes,
-                 show.legend = show.legend
-                 )
+
+  ggplot2::layer(
+    geom = ggplot2::GeomSmooth, stat = StatCategoricalModel,
+    data = data, mapping = mapping, position = position,
+    params = params, inherit.aes = inherit.aes,
+    show.legend = show.legend
+  )
 }
 
 StatCategoricalModel <- ggplot2::ggproto("StatCategoricalModel", ggplot2::Stat,
-  
   required_aes = c("x", "y"),
-  
+
   compute_panel = function(data, scales, se = TRUE, level = .95) {
     if (nrow(data) == 0) {
       return(data[integer(0), ])
     }
-    
+
     if (!scales$x$is_discrete()) {
       warning("`geom_categorical_model()` only works with a discrete x axis variable",
-              call. = FALSE
+        call. = FALSE
       )
       return(data[integer(0), ])
     }
-    
+
     data$x <- factor(data$x)
-    
+
     # Fit model
     model <- stats::lm(formula = y ~ x, data = data)
-    
+
     # Compute prediction from model based on sequence of x-values defined for
[…]
