+++ tests/testthat/test_geom_categorical_model.R
@@ -2,10 +2,10 @@ context("test-geom_categorical_model")
 
 library(ggplot2)
 
-mpg <- mpg[mpg$cyl != 5,]
+mpg <- mpg[mpg$cyl != 5, ]
 
-viz <- ggplot(mpg, aes(x=drv, y=hwy)) +
- geom_point()
+viz <- ggplot(mpg, aes(x = drv, y = hwy)) +
+  geom_point()
 
 
 # geom_categorical_model ----------------------------------------------------
@@ -13,24 +13,24 @@ test_that("geom_categorical_model works", {
   vdiffr::expect_doppelganger(
     "geom_categorical_model-basic",
     viz + geom_categorical_model() + labs(title = "geom_categorical_model()")
-    )
+  )
 
   vdiffr::expect_doppelganger(
     "geom_categorical_model-no-SE",
     viz + geom_categorical_model(se = FALSE) +
       labs(title = "geom_categorical_model() with `se = FALSE`")
-    )
+  )
 
   vdiffr::expect_doppelganger(
     "geom_categorical_model-basic-color-and-size",
-    viz %+% aes(color=drv) +
+    viz %+% aes(color = drv) +
       geom_categorical_model(size = 3) +
       labs(title = "geom_categorical_model() with extra aesthetics")
-    )
-  
+  )
+
   vdiffr::expect_doppelganger(
     "geom_categorical_model-linetype-override",
-    viz %+% aes(linetype=drv) +
+    viz %+% aes(linetype = drv) +
       geom_categorical_model() +
       labs(title = "geom_categorical_model() with linetype mapped")
   )
@@ -40,9 +40,8 @@ test_that("geom_categorical_model works", {
     viz + geom_categorical_model() +
       facet_wrap(~cyl) +
       labs(title = "Faceted geom_categorical_model()")
-    )
-  }
-)
+  )
+})
 
 test_that("geom_categorical_model works in edge cases", {
 
@@ -50,7 +49,7 @@ test_that("geom_categorical_model works in edge cases", {
   expect_warning(
     vdiffr::expect_doppelganger(
       "geom_parallel_slopes-numeric-x",
-      ggplot(mpg, aes(x=displ, y=hwy)) +
+      ggplot(mpg, aes(x = displ, y = hwy)) +
         geom_point() +
         geom_categorical_model() +
         labs(title = "geom_categorical_model() does nothing with numeric x")
