+++ tests/testthat/test-geom_parallel_slopes.R
@@ -75,17 +75,17 @@ test_that("geom_parallel_slopes works", {
       geom_parallel_slopes(aes(fill = gr)) +
       labs(title = "geom_parallel_slopes() with `fill` aesthetics grouping")
   )
-  
+
   vdiffr::expect_doppelganger(
     "geom_parallel_slopes-fullrange",
-    viz + geom_parallel_slopes(fullrange=TRUE) +
-      xlim(c(-1,2)) +
+    viz + geom_parallel_slopes(fullrange = TRUE) +
+      xlim(c(-1, 2)) +
       labs(title = "geom_parallel_slopes() with fullrange=TRUE")
   )
-  
+
   vdiffr::expect_doppelganger(
     "geom_parallel_slopes-level",
-    viz + geom_parallel_slopes(level=0.25) +
+    viz + geom_parallel_slopes(level = 0.25) +
       labs(title = "geom_parallel_slopes() with level=0.25")
   )
 })
