### 2020-06-29 83641d5
add deprecate warning message and update NEWS.md
