### 2020-06-29 8b8e5f6
Tweaked gg_parallel_slopes description
