+++ R/ggplot_parallel_slopes.R
@@ -4,8 +4,8 @@ globalVariables(c(
 
 #' Plot parallel slopes model
 #'
-#' NOTE: THIS FUNCTION WILL SOON BE DEPRECATED; PLEASE USE \code{{geom_parallel_slopes()}}
-#' INSTEAD. Output a visualization of linear regression when you have one numerical
+#' NOTE: This function is deprecated; please use \code{\link{geom_parallel_slopes}}
+#' instead. Output a visualization of linear regression when you have one numerical
 #' and one categorical explanatory/predictor variable: a separate colored
 #' regression line for each level of the categorical variable
 #'
