+++ man/gg_parallel_slopes.Rd
@@ -27,8 +27,8 @@ in \code{data}}
 A \code{\link[ggplot2]{ggplot}} object.
 }
 \description{
-NOTE: THIS FUNCTION WILL SOON BE DEPRECATED; PLEASE USE \code{{geom_parallel_slopes()}}
-INSTEAD. Output a visualization of linear regression when you have one numerical
+NOTE: This function is deprecated; please use \code{\link{geom_parallel_slopes}}
+instead. Output a visualization of linear regression when you have one numerical
 and one categorical explanatory/predictor variable: a separate colored
 regression line for each level of the categorical variable
 }
