### 2020-06-29 bd1480b
Set gg_parallel_slopes examples to dontrun
