+++ man/gg_parallel_slopes.Rd
@@ -33,6 +33,7 @@ and one categorical explanatory/predictor variable: a separate colored
 regression line for each level of the categorical variable
 }
 \examples{
+\dontrun{
 library(ggplot2)
 library(dplyr)
 library(moderndive)
@@ -62,6 +63,7 @@ ggplot(house_prices, aes(x = log10_size, y = log10_price, col = condition)) +
     x = "log10 square feet living space", y = "log10 price in USD",
     title = "House prices in Seattle: Interaction model"
   )
+}
 }
 \seealso{
 \code{\link{geom_parallel_slopes}}
