+++ NEWS.md
@@ -1,14 +1,15 @@
+# moderndive 0.5.0
 
-
-# moderndive 0.4.0.9000
-
-* Modified `vignettes/why-moderndive.Rmd` to act as submission to [Journal of Open Source Education](https://jose.theoj.org/) 
-* "Why ModernDive?" vignette now loads `patchwork` package from CRAN
+* "Why ModernDive?" vignette
+    + Modified `vignettes/why-moderndive.Rmd` to also output in `rticles` format
+    + Now loads `patchwork` package from CRAN
+* Add deprecation warning message to `gg_parallel_slopes()` to use `geom_parallel_slopes()` instead
 * Update `geom_parallel_slopes()` with new arguments:
     + Use `fullrange=TRUE` to draw regression lines over the entire support of the x-axis (by @wjhopper)
     + Use `level` to set different level of confidence interval shading (by @echasnovski)
 * Added new function `geom_categorical_model()` for visualizing regression models with one categorical explanatory/predictor variable (by @wjhopper)
-* Add deprecation warning message to `gg_parallel_slopes()` to use `geom_parallel_slopes()` instead
+
+
 
 ***
 
