### 2020-06-29 e5838bf
Moved rticles vignette out for now
