+++ DESCRIPTION
@@ -3,8 +3,8 @@ Type: Package
 Title: Tidyverse-Friendly Introductory Linear Regression
 Version: 0.4.0.9000
 Authors@R: c(
-    person("Albert Y.", "Kim", email = "albert.ys.kim@gmail.com", role = c("aut", "cre")),
-    person("Chester", "Ismay", email = "chester.ismay@gmail.com", role = "aut"),
+    person("Albert Y.", "Kim", email = "albert.ys.kim@gmail.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0001-7824-306X")),
+    person("Chester", "Ismay", email = "chester.ismay@gmail.com", role = "aut", comment = c(ORCID = "0000-0003-2820-2547")),
     person("Andrew", "Bray", email = "abray@reed.edu", role = "ctb"),
     person("Delaney", "Moran", email = "dmoran@middlebury.edu", role = "ctb"),
     person("Evgeni", "Chasnovski", email = "evgeni.chasnovski@gmail.com", role = "ctb"),
