### 2020-07-13 0a0c7c9
Update testthat documentation
