+++ tests/testthat/test-get_regression_functions.R
@@ -20,7 +20,7 @@ species_glm <- glm(virginica ~ Sepal.Width, data = iris_binary)
 
 test_that("function inputs are valid", {
   vec <- 1:10
-
+  
   # Check `model`
   expect_error(
     get_regression_table(model = vec)
@@ -28,12 +28,12 @@ test_that("function inputs are valid", {
   expect_error(
     get_regression_summaries(model = species_glm)
   )
-
+  
   # Check `digits`
   expect_error(
     get_regression_table(model = mpg_null, digits = "blah")
   )
-
+  
   # Check `print`
   expect_error(
     get_regression_points(model = mpg_cyl, digits = 5, print = "yes")
@@ -41,7 +41,7 @@ test_that("function inputs are valid", {
   expect_silent(
     get_regression_points(model = mpg_cyl, digits = 4, print = TRUE)
   )
-
+  
   # Check newdata
   expect_silent(
     get_regression_points(model = mpg_cyl, ID = "automobile", newdata = mtcars)
@@ -52,7 +52,7 @@ test_that("function inputs are valid", {
   expect_error(
     get_regression_points(model = mpg_cyl, ID = "automobile", newdata = matrix(mtcars))
   )
-
+  
   # Check ID
   expect_error(
     get_regression_points(model = mpg_cyl, ID = 6)
@@ -65,7 +65,7 @@ test_that("README code works", {
   mpg_model <- lm(mpg ~ hp, data = mtcars)
   mpg_mlr_model <- lm(mpg ~ hp + wt, data = mtcars)
   mpg_mlr_model2 <- lm(mpg ~ hp + cyl, data = mtcars)
-
+  
   # Regression tables
   expect_silent(
     get_regression_table(model = mpg_model)
@@ -73,25 +73,25 @@ test_that("README code works", {
   expect_silent(
     get_regression_table(mpg_mlr_model, digits = 4, print = TRUE)
   )
-
+  
   # Regression points. For residual analysis for example
   expect_silent(
     get_regression_points(mpg_mlr_model2)
   )
-  # Case when true observed outcome variable is included and hence we can
-  # compute fitted/predicted values and residuals
+  # newdata case when true observed outcome variable is included and hence we can
+  # compute residuals
   newcars <- slice(mtcars, 1:3)
   expect_silent(
     get_regression_points(mpg_mlr_model2, newdata = newcars)
   )
-  # Case when true observed outcome variable is NOT included and hence we
-  # CANNOT compute fitted/predicted values and residuals
+  # newdata case when true observed outcome variable is not included and hence we
+  # cannot compute residuals
   newcars <- slice(mtcars, 1:3) %>%
     select(-mpg)
   expect_silent(
     get_regression_points(mpg_mlr_model2, newdata = newcars)
   )
-
+  
   # Regression summaries
   expect_silent(
     get_regression_summaries(mpg_model)
