+++ R/datasets.R
@@ -160,7 +160,7 @@
 #'   \item{cls_students}{Total number of students in class.}
 #'   \item{cls_level}{Class level: lower, upper.}
 #' }
-#' @source Çetinkaya-Rundel M, Morgan KL, Stangl D. 2013. Looking Good on Course Evaluations. CHANCE 26(2). \url{https://chance.amstat.org/2013/04/looking-good/}
+#' @source Çetinkaya-Rundel M, Morgan KL, Stangl D. 2013. Looking Good on Course Evaluations. CHANCE 26(2). 
 #' @seealso The data in `evals` is a slight modification of \code{\link[openintro]{evals}}.
 #' @examples
 #' library(dplyr)
