### 2020-07-17 4ef4c67
Fixed LICENSE issue
