### 2020-07-17 56e4c65
more license edits
