### 2020-07-17 1d99026
cleanedup vignettes
