+++ vignettes/paper.md
@@ -1,715 +1,0 @@
----
-title: "Take a `moderndive` into introductory linear regression with R"
-author: "Albert Y. Kim, Chester Ismay, and Max Kuhn"
-date: "2020-03-19"
-vignette: >
-  %\VignetteIndexEntry{Why should you use the moderndive package for intro linear regression?}
-  %\VignetteEncoding{UTF-8}
-  %\VignetteEngine{knitr::rmarkdown}
-tags:
-  - R
-  - Rstats
-  - tidyverse
-  - regression
-  - moderndive
-  - modeling
-  - modelling
-  - kaggle
-authors:
-  - name: Albert Y. Kim
-    orcid: 0000-0001-7824-306X
-    affiliation: 1 
-  - name: Chester Ismay
-    orcid: 0000-0003-2820-2547
-    affiliation: 2
-  - name: Max Kuhn
-    orcid: 0000-0003-2402-136X
-    affiliation: 3   
-affiliations:
- - name: Assistant Professor of Statistical and Data Sciences, Smith College, Northampton, MA, USA.
-   index: 1
- - name: Data Science Evangelist, DataRobot, Portland, OR, USA.
-   index: 2
- - name: Software Engineer, RStudio, USA.
-   index: 3
-bibliography: paper.bib
-output:
-  rticles::joss_article:
-    keep_md: yes
-    number_sections: yes
-  rmarkdown::html_vignette:
-    df_print: default
-    number_sections: yes
----
-
-
-
-
-
-
-
-# Introduction
-
-Linear regression has long been a staple of introductory statistics courses. While the curricula of introductory statistics courses has much evolved of late, the overall importance of regression remains the same [@ASAGuidelines]. To facilitate the teaching of regression while leveraging modern computational tools, we've created the [`moderndive`](https://moderndive.github.io/moderndive/){target="_blank"} R package of datasets and functions for [tidyverse](https://www.tidyverse.org/)-friendly introductory linear regression [@tidyverse2019]. 
-
-This R package is designed to supplement the book "Statistical Inference via Data Science: A ModernDive into R and the Tidyverse" [@ismay2019moderndive]. Note that the book is also available online at https://www.moderndive.com and is referred to as "ModernDive" for short.
-
-Before we proceed, let's load all the R packages we are going to need.
-
-
-```r
-library(moderndive)
-library(ggplot2)
-library(dplyr)
-library(readr)
-library(knitr)
-library(broom)
-```
-
-Let's consider data gathered from end of semester student evaluations for a sample of 463 courses taught by 94 professors from the University of Texas at Austin [@diez2015openintro]. This data is included in the `evals` data frame from the `moderndive` package.
-
-
-
-In the following table, we present a subset of 9 of the 14 variables included for a random sample of 5 courses^[For details on the remaining 5 variables, see the help file by running `?evals`.]:
-
-1. `ID` uniquely identifies the course whereas `prof_ID` identifies the professor who taught this course. This distinction is important since many professors taught more than one course.
-1. `score` is the outcome variable of interest: average professor evaluation score out of 5 as given by the students in this course.
-1. The remaining variables are demographic variables describing that course's instructor, including `bty_avg` (average "beauty" score) for that professor as given by a panel of 6 students.^[Note that `gender` was collected as a binary variable at the time of the study (2005).]
-
-
-  ID   prof_ID   score   age   bty_avg  gender   ethnicity      language   rank    
-----  --------  ------  ----  --------  -------  -------------  ---------  --------
- 129        23     3.7    62     3.000  male     not minority   english    tenured 
- 109        19     4.7    46     4.333  female   not minority   english    tenured 
[…]
