### 2020-07-17 3d27338
Preparing cran submission
