+++ cran-comments.md
@@ -3,10 +3,14 @@
 * local macOS install, R 4.0.1
 * ubuntu 16.04.6 (on travis-ci), R-release, R-devel, R-oldrel
 * win-builder (release, devel)
+* GitHub Actions
+    + ubuntu-16.04: release
+    + windows: release
+    + macOS: release, devel
 * Rhub
-    + Windows Server 2008 R2 SP1, R-devel, 32/64 bit
-    + Ubuntu Linux 16.04 LTS, R-release, GCC
     + Fedora Linux, R-devel, clang, gfortran
+    + Ubuntu Linux 16.04 LTS, R-release, GCC
+    + Windows Server 2008 R2 SP1, R-devel, 32/64 bit
 
     
 ## R CMD check results
