### 2020-07-17 2ee8c6b
revdepcheck::revdep_check() & devtools::spell_check() ✅
