### 2020-07-19 bbd7071
Reverting why-moderndive rticles output
