+++ vignettes/paper.md
@@ -1,9 +1,9 @@
 ---
 title: "Take a `moderndive` into introductory linear regression with R"
 author: "Albert Y. Kim, Chester Ismay, and Max Kuhn"
-date: "`r Sys.Date()`"
+date: "2020-03-19"
 vignette: >
-  %\VignetteIndexEntry{Take a `moderndive` into introductory linear regression with R}
+  %\VignetteIndexEntry{Why should you use the moderndive package for intro linear regression?}
   %\VignetteEncoding{UTF-8}
   %\VignetteEngine{knitr::rmarkdown}
 tags:
@@ -34,63 +34,15 @@ affiliations:
    index: 3
 bibliography: paper.bib
 output:
-# Uncomment the four lines below to build JOSE article using rticles:
-#  rticles::joss_article:
-#    keep_md: yes
-#    number_sections: yes
-#    citation_package: default
+  rticles::joss_article:
+    keep_md: yes
+    number_sections: yes
   rmarkdown::html_vignette:
     df_print: default
     number_sections: yes
 ---
 
-```{r, include = FALSE}
-# knitr settings
-knitr::opts_chunk$set(
-  # Code output:
-  warning = FALSE,
-  message = FALSE,
-  echo = TRUE,
-  # Figure:
-  out.width = "100%",
-  fig.path = "Figures/",
-  fig.width = 16 / 2.5,
-  fig.height = 9 / 2.5,
-  fig.align = "center",
-  fig.show = "hold",
-  # Etc:
-  collapse = TRUE,
-  comment = "##"
-  # tidy = FALSE
-)
 
-# Needed packages in vignette
-library(moderndive)
-library(ggplot2)
-library(dplyr)
-library(readr)
-library(knitr)
-library(broom)
-library(viridis)
-
-# Needed packages internally
-library(patchwork)
-
-# Random number generator seed value
-set.seed(76)
-
-# Set ggplot defaults for JOSE output:
-if (!knitr::is_html_output()) {
-  # Grey theme:
-  theme_set(theme_light())
-
-  scale_colour_discrete <- scale_colour_viridis_d
-}
-
-
-# Set output width for JOSE paper:
-options(width = 70)
-```
 
 
 
@@ -104,7 +56,8 @@ This R package is designed to supplement the book "Statistical Inference via Dat
 
 Before we proceed, let's load all the R packages we are going to need.
 
-```{r}
+
+```r
 library(moderndive)
 library(ggplot2)
 library(dplyr)
@@ -115,22 +68,22 @@ library(broom)
 
 Let's consider data gathered from end of semester student evaluations for a sample of 463 courses taught by 94 professors from the University of Texas at Austin [@diez2015openintro]. This data is included in the `evals` data frame from the `moderndive` package.
 
-```{r, echo=FALSE}
-evals_sample <- evals %>%
-  select(ID, prof_ID, score, age, bty_avg, gender, ethnicity, language, rank) %>%
-  sample_n(5)
-```
 
-In the following table, we present a subset of `r ncol(evals_sample)` of the `r ncol(evals)` variables included for a random sample of `r nrow(evals_sample)` courses^[For details on the remaining `r ncol(evals) - ncol(evals_sample)` variables, see the help file by running `?evals`.]:
+
+In the following table, we present a subset of 9 of the 14 variables included for a random sample of 5 courses^[For details on the remaining 5 variables, see the help file by running `?evals`.]:
 
 1. `ID` uniquely identifies the course whereas `prof_ID` identifies the professor who taught this course. This distinction is important since many professors taught more than one course.
 1. `score` is the outcome variable of interest: average professor evaluation score out of 5 as given by the students in this course.
[…]
