### 2020-07-19 4d718e3
Revert "Reverting why-moderndive rticles output"
This reverts commit bbd7071b0db9dccfd3f5dce4cf7fce64f5bb6eb5.
