### 2020-07-19 e60bdf5
Revert "Version bump"
This reverts commit 744f9e9915c2e5058078392d89a972ebd821a7f0.
