+++ NEWS.md
@@ -1,13 +1,3 @@
-# moderndive 0.5.0.9000
-
-* 
-
-
-
-***
-
-
-
 # moderndive 0.5.0
 
 * Modified `vignettes/why-moderndive.Rmd` main vignette
