### 2020-07-19 9e3f0e9
Revert "Revert "Version bump""
This reverts commit e60bdf542306854bc5a6a0993718a28b1990ca8b.
