+++ NEWS.md
@@ -1,3 +1,13 @@
+# moderndive 0.5.0.9000
+
+* 
+
+
+
+***
+
+
+
 # moderndive 0.5.0
 
 * Modified `vignettes/why-moderndive.Rmd` main vignette
