### 2020-07-19 00c9b4c
Revert "Revert "Reverting why-moderndive rticles output""
This reverts commit 4d718e31eebf268e8ef8825ff7b333e8cb727784.
