+++ pkgdown.yml
@@ -1,7 +1,7 @@
-pandoc: 2.3.1
+pandoc: 2.7.3
 pkgdown: 1.5.1
 pkgdown_sha: ~
 articles:
   why-moderndive: why-moderndive.html
-last_built: 2020-06-17T21:27Z
+last_built: 2020-07-19T15:00Z
 
