+++ cran-comments.md
@@ -1,13 +1,13 @@
 ## Test environments
 
 * local macOS install, R 4.0.1
-* ubuntu 16.04.6 (on travis-ci), R-release, R-devel, R-oldrel
-* win-builder (release, oldrelease)
+* ubuntu 16.04.6 on travis-ci (release, devel, oldrel)
+* win-builder (release, devel, oldrel)
 * GitHub Actions
     + ubuntu-16.04: release
     + windows: release
     + macOS: release, devel
-* Rhub
+* Rhub via devtools::check_rhub(env_vars=c(R_COMPILE_AND_INSTALL_PACKAGES = "always"))
     + Fedora Linux, R-devel, clang, gfortran
     + Windows Server 2008 R2 SP1, R-devel, 32/64 bit
 
