### 2020-09-29 #80 @moorepants
@rudeboybert After a quick look at the paper, it seems this is primarily a software library. My initial thought is that this is more appropriate for JOSS than JOSE. We accept two types of submissions as described here: https://jose.theoj.org/about

This doesn't seem to be "open-source education materials" and then this is the description of the type of software we accept:

> What do you mean by "educational software tools"?
>
> Open-source software that serves as educational technology; examples include (but are not limited to) alternatives to learning management systems, autograders, cloud systems for lesson delivery, student collaboration tools. For these tools, peer review will follow a similar process as JOSS.

Can you provide some explanation of which of the two this fits? Thanks.
