### 2020-10-06 #80 @rudeboybert
Hi @moorepants, thanks for your reply. The paper does center around a software library and I agree that by itself it would be a stretch to call this library "open-source education materials". I also agree that this library doesn't satisfy your definition of "educational software tools."

However this library is meant to be used as a complement to an online textbook ["Statistical Inference via Data Science: A ModernDive into R & the Tidyverse"](http://moderndive.com/) (which also has a print-edition). In particular the library provides:

1. Easy access to the datasets used in the book so that novices need minimal data skills to access rich, real, and relevant datasets.
1. More importantly, makes introductory regression more accessible. For example by providing wrapper functions for unpacking regression results and tools for data visualization.

Given that (1) this library is an important component of our ecosystem that facilitates lessons centered on computation and (2) the contents of the paper can also be used as a standalone refresher on regression, I feel that this is an appropriate submission for JOSE under the "open-source education materials" rubric.

Please let me know what you think.
