### 2020-10-06 #80 @moorepants
@rudeboybert Thanks for your explanation. As I understand it, this wouldn't fall under "open source education materials" unless you have content that is making use of this library. It sounds like you may already have published the content in the text book. I'd like some of the other @openjournals/jose-editors weigh in on this, as I've been confused about this in the past.
