### 2020-10-06 #80 @moorepants
I just browsed the online textbook. That is exactly the type of material we want to publish in JOSE. So why don't you submit that?
