### 2020-10-06 #80 @moorepants
> So why don't you submit that?

I may see the answer to my question. You've published this with CRC press already.
