### 2021-01-05 #80 @rudeboybert
Hi @moorepants, I hope you've been well. I was wondering if you had a response about the suitably of this manuscript for JOSE. To clarify, while this package is written as a companion to our book that has been published by CRC press, it can act as a standalone package to facilitate teaching introductory regression.
