### 2021-01-05 ec3963a
Use vdiffr conditionally
