+++ NEWS.md
@@ -1,6 +1,6 @@
-# moderndive 0.5.0.9000
+# moderndive 0.5.1
 
-* 
+* Use vdiffr conditionally
 
 
 
