+++ DESCRIPTION
@@ -34,7 +34,8 @@ Imports:
     rlang (>= 0.2.0),
     glue
 RoxygenNote: 7.1.1
-Suggests: testthat,
+Suggests: usethis,
+    testthat,
     covr,
     rmarkdown,
     vdiffr,
