### 2021-01-07 5cabdc8
Updated github actions, removed travis
