+++ .travis.yml
@@ -1,46 +1,0 @@
-# Sample .travis.yml for R projects
-
-language: r
-
-matrix:
-  include:
-  - r: oldrel
-    # Don't use vdiffr on old R release
-    env: VDIFFR_RUN_TESTS=false
-  - r: release
-  - r: devel
-
-warnings_are_errors: true
-cache: packages
-latex: false
-
-env:
- global:
-   - CRAN: http://cran.rstudio.com
-   # Use vdiffr by default
-   - VDIFFR_RUN_TESTS: true
-
-notifications:
-  email:
-    on_success: change
-    on_failure: change
-
-r_packages:
-  - devtools
-  
-r_github_packages:
-  - hadley/pkgdown
-  - moderndive/moderndive
-
-before_deploy:
-  - Rscript -e 'pkgdown::build_site()'
-
-deploy:
-  provider: pages
-  cleanup: true
-  keep-history: true
-  local-dir: docs
-  github_token: $MODERNDIVE_GITHUB_PAT
-
-after_success:
-  - Rscript -e 'covr::codecov()'
