+++ .github/workflows/check-standard.yaml
@@ -25,6 +25,8 @@ jobs:
 
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
+      # don't treat missing suggested packages as error:
+      _R_CHECK_FORCE_SUGGESTS_: false
       RSPM: ${{ matrix.config.rspm }}
 
     steps:
