### 2021-01-08 7c260af
update CRAN comments
