+++ NEWS.md
@@ -1,3 +1,13 @@
+# moderndive 0.5.1.9000
+
+* 
+
+
+
+***
+
+
+
 # moderndive 0.5.1
 
 * Use vdiffr conditionally
