### 2021-02-01 #80 @labarba
Hi everybody 👋
Thank you all for your patience. We've all been pretty deluged (and @moorepants changed countries during a pandemic!).

We often struggle with scope questions, as JOSE is an unusual journal. If I understand correctly, this software was written to _accompany_ a textbook, and thus it has an educational goal. Even if we did not consider this type of submission, I do believe we should consider it in-scope. The book is published as static text content. The software provides computational learning experiences. Is that the idea?

BTW, @moorepants, if you're deluged right now, we have a new R expert in the editorial who could help us out here: @ttimbers 🙌
