### 2021-02-01 #80 @rudeboybert
Hello @labarba, many thanks for your response.

To answer your question: yes, the package provides computational learning experiences (while the book is published as static text content at moderndive.com). The idea being that while recent developments in R and the tidyverse have made linear regression performed on a computer much more accessible to the masses, our software package provides a few extra tools to bridge some remaining gaps. The manuscript showcases these tools with exercises/activities that can be used in the classroom/lab.

Thank you for the energy you and your team put into JOSE.
