### 2021-02-11 52ba67d
Update get_correlation.R
The explanatory variables are on the _right_ hand side of the equation, not the left.
