### 2021-02-20 8719980
add extract_cat_names function and tests
