### 2021-02-22 b4e65d3
Fix GitHub actions per @patrickvossler18 suggestion
