+++ .github/workflows/check-standard.yaml
@@ -46,7 +46,7 @@ jobs:
       
       - name: Install XQuartz on MacOS
         if: runner.os == 'macOS'
-        run: brew cask install xquartz
+        run: brew install --cask xquartz
 
       - name: Install freetype on MacOS
         if: runner.os == 'macOS'
