### 2021-02-23 9d4c427
add boolean argument to allow for default behavior
also added tests for this argument
