### 2021-02-23 c134799
update documentation for get_regression_table
Include the default_categorical_levels boolean argument
