### 2021-02-23 8510ddc
Update NEWS.md
