### 2021-02-25 c155638
Update NEWS.md
