+++ NEWS.md
@@ -4,17 +4,9 @@
 
 
 
-***
-
-
-
 # moderndive 0.5.1
 
 * Use vdiffr conditionally
-
-
-
-***
 
 
 
@@ -29,10 +21,6 @@
 
 
 
-***
-
-
-
 # moderndive 0.4.0
 
 * Added `geom_parallel_slopes()` geom extension to `ggplot2` package to plot parallel slopes regression models with one numerical and one categorical variable (this is not possible using `ggplot2::geom_smooth()`). Note this renders `gg_parallel_slopes()` function added in v0.3.0 obsolete.
@@ -41,10 +29,6 @@
 * `get_correlation()` now:
     + Respects `dplyr::group_by()` grouping
     + Can handle missing data by either passing `na.rm = TRUE` argument or by passing standard `stats:cor(use = "complete.obs")` argument via `...`
-
-
-
-***
 
 
 
@@ -68,10 +52,6 @@ observational units/rows
 
 
 
-***
-
-
-
 # moderndive 0.2.0
 
 Updated package for:
@@ -91,15 +71,11 @@ Details:
 * Removed `tidyverse` from Depends, Imports, or Suggests
 
 
-***
-
 
 # moderndive 0.1.1
 
 Fixed broken url in `?bowl_samples`
 
-
-***
 
 
 # moderndive 0.1.0
