### 2021-03-03 29e837c
change extract_cat_names to handle in-line factors
Also add test condition to check that this works properly
