### 2021-03-03 a754068
add additional case of factor(cyl)
