### 2021-04-21 #80 @moorepants
@ttimbers, this submission fell off my radar. @rudeboybert wrote me today to ask about the status. I'm not an R or stats expert. Given that Lorena suggested you, might you be interested in editing this submission? If not, maybe you have some suggestions on reviewers? Thanks -- Jason
