### 2021-04-28 #80 @moorepants
@gvdr, @kbodwin, @lebebr01, @jhollist, we are looking for reviewers fro this JOSE submission entitled "Take a moderndive into introductory linear regression with R". Would any of you like to review this submission?
