### 2021-04-28 #80 @jhollist
Sorry, but I am going to have to pass.  I am swamped with reviews at the
moment.

Keep me in mind for the future though!

On Wed, Apr 28, 2021 at 3:39 PM Jason K. Moore ***@***.***>
wrote:

> @gvdr <https://github.com/gvdr>, @kbodwin <https://github.com/kbodwin>,
> @lebebr01 <https://github.com/lebebr01>, @jhollist
> <https://github.com/jhollist>, we are looking for reviewers fro this JOSE
> submission entitled "Take a moderndive into introductory linear regression
> with R". Would any of you like to review this submission?
>
