### 2021-04-28 #80 @moorepants
@jhollist, thanks for the response.

@kbodwin, early mid June may work. I'll see if we can find two that can start earlier and keep you as a backup if we can't. Does that sound fine?
