### 2021-05-06 #80 @moorepants
@lwjohnst86 We are looking for a second reviewer for this Journal of Open Source Education submission "Take a moderndive into introductory linear regression with R". Would you be interested in reviewing it?
