### 2021-05-06 #80 @lwjohnst86
@moorepants for sure! I use that resource myself as inspiration, so I'd love to formally review it. I probably would need a month _at the least_, given my time pressures right now and the size of the material.
