### 2021-05-06 #80 @moorepants
>  I probably would need a month at the least, given my time pressures right now and the size of the material.

Note that this is not a review of the book, but only the accompanying software package and associated documentation.
