### 2021-05-06 #80 @rudeboybert
@moorepants only update it from v0.5.0 to v0.5.1 (latest) if the work involved is minimal. The difference doesn't affect the manuscript: https://github.com/moderndive/moderndive/blob/master/NEWS.md
