### 2021-05-06 #80 @rudeboybert
Ah my mistake, I saw that but somehow read it as v0.5.0. In any case, I just tested the manuscript on the Whedon paper preview service and it rendered correctly.
