### 2021-05-06 #115 @whedon
[REVIEW]: Take a moderndive into introductory linear regression with R
**Submitting author:** @rudeboybert (<a href="http://orcid.org/0000-0001-7824-306X">Albert Y. Kim</a>)
**Repository:** <a href="https://github.com/moderndive/moderndive" target ="_blank">https://github.com/moderndive/moderndive</a>
**Version:** v0.5.2

✨ **Please try and complete your review in the next four weeks** ✨
