### 2021-05-06 #115 @moorepants
@lisamr & @lwjohnst86, the review can now officially start here. Please read from the top of the thread and be sure to review the reviewer guidelines. You'll work your way through the checklists above over the next 4 weeks. Post comments or issues to the main repository as you go along or all at once, your preference. An automatic notice will be sent out in 2 weeks as a reminder. You should review the paper, software documentation, and test out & use the software. The paper authors can (and are encouraged to) interact while you do the review. Let me know if there are any questions. Welcome to JOSE!
