### 2021-05-13 #115 @lisamr
@moorepants I tried to follow this link: "Be sure to accept the invite at this URL: https://github.com/openjournals/joss-reviews/invitations", but I get this message: "Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account." I'm assuming that's where I can edit the checklist, correct?
