### 2021-05-14 #115 @lisamr
I started going through the checklist pasted at the top of this thread:

# General checks
- authorship: It is currently unclear how Max Kuhn contributed, as they are not listed as a contributor on the Github repository and there isn't an author contribution section in the manuscript to clarify.

# Documentation
- Statement of need: It's missing in the Readme file and in the MS
- Installation instructions: I couldn't find a clearly stated list of dependencies in the repository, but maybe I was looking in the wrong place. I checked under the "dependency graph" tab by clicking on the "insight" button, and no dependencies were listed there either.

## Minor details on the readme file
 - Under Demo, it's stated

> Among the many useful features of the moderndive package outlined in our essay “Why should you use the moderndive package for intro linear regression?” we highlight three functions in particular as covered there. We also mention the geom_parallel_slopes() function as #4.

Just say there are 4 functions covered. I'm not sure why `geom_parallel_slopes` is singled out.
- for the function `get_regression_table()` I suggest at least in the help documentation, it should be stated what the alpha level is. I know it's probably 0.05, but it should be stated, or even allowed to vary since the point of the function is to understand confidence intervals between a given range.
- for the function `get_regression_points()`: If this package is aimed at students very new to statistics, it might be helpful to include the following in the readme.md file: "...including fitted/predicted values (e.g. 'score_hat')"
- for the function `get_regression_summaries()`: excuse me if I'm wrong since for the life of me, I never remembered how to calculate degrees of freedom, but should `df = 2`? The output from `summary(score_model)` shows 1 and 461 DF.
