### 2021-05-14 #115 @moorepants
> @moorepants I tried to follow this link: "Be sure to accept the invite at this URL: https://github.com/openjournals/joss-reviews/invitations", but I get this message: "Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account." I'm assuming that's where I can edit the checklist, correct?

@labarba, do the reviewers have to follow this link? It seems like a joss thing that hasnt' converted to jose.
