### 2021-05-14 #115 @moorepants
> Functional documentation and tests: I'm not qualified to assess this since I don't write packages (I'm more of a user than developer)

@lisamr, is it possible for your to try running the tests? It's good if each reviewer can do so. You don't have to inspect the code for them.
