### 2021-05-14 #115 @moorepants
Also, each reviewer should be able to have a look at this too:

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

This is checking to see if there are understandable vignettes for each public function, class, etc.
