### 2021-05-14 #115 @moorepants
Actually, vignettes may have not been precise enough. There is object documentation, see:

https://r-pkgs.org/man.html#man

and vignettes:

https://r-pkgs.org/vignettes.html#vignettes

Ideally, the package will have both of these types of documentation. The first is the "API method documentation" mentioned in the checklist. The second is a bit more long form explanations that aren't tied to a specific object. The object documentation should be present for each public function and the vignettes should be present as per the reviewers' suggestions, mostly to conform with best practices in R documentation.
