### 2021-05-14 #115 @rudeboybert
Hi @lisamr thank you for bringing up those checklist items. I'll make the necessary changes and have Whedon generate a new pdf today.
