### 2021-05-14 f74737d
Added author contributions section
