### 2021-05-14 1803111
Fixed typos in README
