+++ README.Rmd
@@ -10,7 +10,6 @@ knitr::opts_chunk$set(echo = TRUE, fig.path = "man/figures/")
 
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/moderndive)](https://cran.r-project.org/package=moderndive) 
 [![Lifecycle: stable](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://www.tidyverse.org/lifecycle/#stable)
-[![Travis-CI Build Status](https://travis-ci.org/moderndive/moderndive.svg?branch=master)](https://travis-ci.org/moderndive/moderndive) 
 [![GitHub Actions Status](https://github.com/moderndive/moderndive/workflows/R-CMD-check/badge.svg)](https://github.com/moderndive/moderndive/actions)
 [![Coverage Status](https://img.shields.io/codecov/c/github/moderndive/moderndive/master.svg)](https://codecov.io/github/moderndive/moderndive?branch=master)
 [![CRAN RStudio mirror downloads](http://cranlogs.r-pkg.org/badges/moderndive)](http://www.r-pkg.org/pkg/moderndive)
@@ -45,7 +44,7 @@ library(moderndive)
 score_model <- lm(score ~ age, data = evals)
 ```
 
-Among the many useful features of the `moderndive` package outlined in our essay ["Why should you use the moderndive package for intro linear regression?"](https://moderndive.github.io/moderndive/articles/why-moderndive.html) we highlight three functions in particular as covered there. 
+Among the many useful features of the `moderndive` package outlined in our essay ["Why should you use the moderndive package for intro linear regression?"](https://moderndive.github.io/moderndive/articles/why-moderndive.html) we highlight four functions in particular as covered there. 
 
 We also mention the `geom_parallel_slopes()` function as **#4**.
 
