+++ README.md
@@ -13,8 +13,8 @@ downloads](http://cranlogs.r-pkg.org/badges/moderndive)](http://www.r-pkg.org/pk
 
 An R package of datasets and wrapper functions for
 [tidyverse](https://www.tidyverse.org/)-friendly introductory linear
-regression used in "Statistical Inference via Data Science: A ModernDive
-into R and the Tidyverse"" available at
+regression used in “Statistical Inference via Data Science: A ModernDive
+into R and the Tidyverse” available at
 [ModernDive.com](https://moderndive.com/).
 
 ## Installation
@@ -47,7 +47,7 @@ score_model <- lm(score ~ age, data = evals)
 Among the many useful features of the `moderndive` package outlined in
 our essay [“Why should you use the moderndive package for intro linear
 regression?”](https://moderndive.github.io/moderndive/articles/why-moderndive.html)
-we highlight three functions in particular as covered there.
+we highlight four functions in particular as covered there.
 
 We also mention the `geom_parallel_slopes()` function as **\#4**.
 
@@ -100,10 +100,10 @@ mean-squared error:
 get_regression_summaries(score_model)
 ```
 
-    ## # A tibble: 1 x 8
-    ##   r_squared adj_r_squared   mse  rmse sigma statistic p_value    df
-    ##       <dbl>         <dbl> <dbl> <dbl> <dbl>     <dbl>   <dbl> <dbl>
-    ## 1     0.011         0.009 0.292 0.540 0.541      5.34   0.021     2
+    ## # A tibble: 1 x 9
+    ##   r_squared adj_r_squared   mse  rmse sigma statistic p_value    df  nobs
+    ##       <dbl>         <dbl> <dbl> <dbl> <dbl>     <dbl>   <dbl> <dbl> <dbl>
+    ## 1     0.011         0.009 0.292 0.540 0.541      5.34   0.021     1   463
 
 #### 4\. Plot parallel slopes models
 
@@ -118,15 +118,14 @@ ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
   geom_parallel_slopes(se = FALSE)
 ```
 
-![](man/figures/plot-example-1.png)<!-- -->
+<img src="man/figures/plot-example-1.png" width="672" />
 
 ## Other features
 
 #### 1\. Print markdown friendly tables
 
 Want to output cleanly formatted tables in an R Markdown document? Just
-add `print = TRUE` to any of the three `get_regression_*()`
-functions.
+add `print = TRUE` to any of the three `get_regression_*()` functions.
 
 ``` r
 get_regression_table(score_model, print = TRUE)
@@ -170,8 +169,7 @@ write_csv(submission, "submission.csv")
 
 The resulting `submission.csv` is formatted such that it can be
 submitted on Kaggle, resulting in a “root mean squared logarithmic
-error” leaderboard score of
-0.42918.
+error” leaderboard score of 0.42918.
 
 ![](https://github.com/moderndive/moderndive/raw/master/vignettes/leaderboard_orig.png)<!-- -->
 
@@ -226,19 +224,19 @@ library(broom)
 augment(score_model)
 ```
 
-    ## # A tibble: 463 x 9
-    ##    score   age .fitted .se.fit  .resid    .hat .sigma   .cooksd .std.resid
-    ##    <dbl> <int>   <dbl>   <dbl>   <dbl>   <dbl>  <dbl>     <dbl>      <dbl>
[…]
