### 2021-05-14 12a972f
Added conf.level argument to get_regression_table()
