### 2021-05-14 #115 @lisamr
@moorepants got it, so the "functionality documentation" checklist item is referring to the object documentation. @rudeboybert below is my feedback:
- documentation for various datasets look great--good detail on explaining each column and providing source of the data
- `geom_categorical_model()`: It is not obvious how the solid and dashed lines differ or how they were determined and it's not clearly stated in the documentation. I thought maybe pairwise contrasts were being performed, but when I look at two categories that are not significantly different (e.g. `ggplot(evals, aes(pic_outfit, score )) + geom_point() + geom_categorical_model()`), their lines are still different. Then I thought maybe each category gets its own linetype, but when I look at `ggplot(mpg, aes(x = drv, y = hwy)) + geom_point() + geom_categorical_model()`, again there are only 2 types.
- `get_regression_summaries()`: All of the columns are self-explanatory except for df. More details under the "Value" section would be nice.
- `get_regression_table()`: Again, either state that the upper and lower bounds represent the 95% confidence interval, or even better, allow the interval ranges to vary, which in my opinion will help break down this (artificial) idea that 95% CI and the associated p-values are all that matter.

# Larger structural comments on the paper
I went through the paper last night and my first impression was that it was long (~2500 words of text) and overlapped quite a lot with the text and code in the readme file. From what I understand, "Detailed documentation should be present in the repository of the submitted software or module, is reviewed there, and does not need to be included in the paper" (link [here](https://openjournals.readthedocs.io/en/jose/review_criteria.html)). It also lacked a Statement of Need section. I thought a lot of the information presented could have been summed up in a shorter summary paragraph. I also thought the code/output could have been condensed by directing the reader to the repository's documentation.

Here are some key descriptions of the paper, quoted from https://openjournals.readthedocs.io/en/jose/submitting.html:
- Write a short Markdown file titled paper.md with title, author names and affiliations, containing a description of the software, statement of need, and key references.
- JOSE papers contain a limited set of metadata, plus Summary & Reference sections. We explicitly do not publish long-form articles, because the scholarship represented by a JOSE article is contained in the software or learning modules themselves. Expected length is around **1000 words max**.

@moorepants lmk if I'm misunderstanding anything or maybe clarify how the software paper should differ from the readme and other vignettes?
