### 2021-05-14 #115 @moorepants
> @moorepants lmk if I'm misunderstanding anything or maybe clarify how the software paper should differ from the readme and other vignettes?

Yes, you've interpreted correctly. The documentation and paper should be different as you've described.
