### 2021-05-14 699d503
Added conf.level = 0.99 example to vignette
