### 2021-05-14 #115 @rudeboybert
Hi, I had almost finished addressing the initial comments by @lisamr when they followed up with more feedback. I will address the initial comments only here:

> I started going through the checklist pasted at the top of this thread:
>
> # General checks
> * authorship: It is currently unclear how Max Kuhn contributed, as they are not listed as a contributor on the Github repository and there isn't an author contribution section in the manuscript to clarify.

I added a "Author contributions" section towards the end of the paper.
> # Documentation
> * Statement of need: It's missing in the Readme file and in the MS

A "Statement of need" section has been added at the outset of the manuscript.
> * Installation instructions: I couldn't find a clearly stated list of dependencies in the repository, but maybe I was looking in the wrong place. I checked under the "dependency graph" tab by clicking on the "insight" button, and no dependencies were listed there either. https://github.com/moderndive/moderndive/network/dependencies

All the dependencies are automatically installed when installing the `moderndive` package, the list of all dependencies can be seen in the [`DESCRIPTION`](https://github.com/moderndive/moderndive/blob/master/DESCRIPTION) file
> * Functional documentation and tests: I'm not qualified to assess this since I don't write packages (I'm more of a user than developer). @lwjohnst86 do you want to make sure to cover that component?

@moorepants addressed this above
>
> ## Minor details on the readme file
> * Under Demo, it's stated
>
> > Among the many useful features of the moderndive package outlined in our essay “Why should you use the moderndive package for intro linear regression?” we highlight three functions in particular as covered there. We also mention the geom_parallel_slopes() function as #4.
>
> Just say there are 4 functions covered. I'm not sure why `geom_parallel_slopes` is singled out.

That was a typo: The README has been updated to indicate that there are 4 functions highlighted.
>
> * for the function `get_regression_table()` I suggest at least in the help documentation, it should be stated what the alpha level is. I know it's probably 0.05, but it should be stated, or even allowed to vary since the point of the function is to understand confidence intervals between a given range.

Thank you for this great suggestion. We updated the functionality of this function to allow for a `conf.level` argument in this [commit](https://github.com/moderndive/moderndive/pull/104/commits/12a972f48f9d092475be5cbebf2ac9956c807e96), and updated the README and manuscript accordingly. This is only available in the development version of the package `v0.5.1.9000` for now, I'll wait for more feedback before updating on CRAN. @moorepants could you update the version to `v0.5.1.9000`?
> * for the function `get_regression_points()`: If this package is aimed at students very new to statistics, it might be helpful to include the following in the readme.md file: "...including fitted/predicted values (e.g. 'score_hat')"

Thank you, similar language was added to the README in this [commit](https://github.com/moderndive/moderndive/pull/104/commits/f0a775d7758f17ba0c538c44ab309485c1f79a6f)
> * for the function `get_regression_summaries()`: excuse me if I'm wrong, but should `df = 2`? The output from `summary(score_model)` shows 1 and 461 DF.

You are correct. This seems to have been a bug in the `broom::glance()` function that has since been fixed; the README has been updated.
