### 2021-05-21 #115 @lisamr
> 👋 @lisamr, please update us on how your review is going (this is an automated reminder).

I've gone through the checklist as much as I can and I'm actually still waiting on the invite to this URL: https://github.com/openjournals/joss-reviews/invitations. The next major steps are to review the newly shortened version of the paper.
