### 2021-05-21 #115 @moorepants
> I'm actually still waiting on the invite to this URL: https://github.com/openjournals/joss-reviews/invitations.

@lisamr I don't think this is important. Don't worry about it. The link is broken as it is for joss, not jose.
