### 2021-05-21 #115 @lisamr
> > I've gone through the checklist as much as I can
>
> You can check the boxes at the top for the things you've approved as complete.

Hmm, I'm afraid I don't know how to edit the checklist up top. Help?
