### 2021-05-21 #115 @moorepants
> Hmm, I'm afraid I don't know how to edit the checklist up top. Help?

Presumably when you accept the new link invite you'll be able to check the boxes.
