### 2021-05-21 #115 @labarba
@openjournals/dev Help!
We have a reviewer unable to edit the checkboxes. Can you check the issue?
