+++ DESCRIPTION
@@ -34,7 +34,8 @@ Imports:
     rlang (>= 0.2.0),
     glue
 RoxygenNote: 7.1.1
-Suggests: testthat,
+Suggests: 
+    testthat,
     covr,
     rmarkdown,
     vdiffr,
@@ -42,5 +43,6 @@ Suggests: testthat,
     openintro,
     patchwork,
     viridis,
-    readr
+    readr,
+    ragg
 VignetteBuilder: knitr
