+++ _pkgdown.yml
@@ -1,0 +1,1 @@
+url: https://moderndive.github.io/moderndive/
