### 2021-06-04 788c541
update codecov yml based on use_coverage().
