+++ README.md
@@ -6,8 +6,8 @@
 stable](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://www.tidyverse.org/lifecycle/#stable)
 [![GitHub Actions
 Status](https://github.com/moderndive/moderndive/workflows/R-CMD-check/badge.svg)](https://github.com/moderndive/moderndive/actions)
-[![Coverage
-Status](https://img.shields.io/codecov/c/github/moderndive/moderndive/master.svg)](https://codecov.io/github/moderndive/moderndive?branch=master)
+[![Codecov test
+coverage](https://codecov.io/gh/moderndive/moderndive/branch/master/graph/badge.svg)](https://codecov.io/gh/moderndive/moderndive?branch=master)
 [![CRAN RStudio mirror
 downloads](http://cranlogs.r-pkg.org/badges/moderndive)](http://www.r-pkg.org/pkg/moderndive)
 
@@ -51,7 +51,7 @@ we highlight four functions in particular as covered there.
 
 We also mention the `geom_parallel_slopes()` function as **\#4**.
 
-#### 1\. Get regression tables
+#### 1. Get regression tables
 
 Get a tidy regression table **with** confidence intervals:
 
@@ -78,7 +78,7 @@ get_regression_table(score_model, conf.level = 0.99)
     ## 1 intercept    4.46      0.127     35.2    0        4.13     4.79 
     ## 2 age         -0.006     0.003     -2.31   0.021   -0.013    0.001
 
-#### 2\. Get fitted/predicted values and residuals
+#### 2. Get fitted/predicted values and residuals
 
 Get information on each point/observation in your regression, including
 fitted/predicted values & residuals (e.g. `score_hat` and `residual`),
@@ -103,7 +103,7 @@ get_regression_points(score_model)
     ## 10    10   4.5    40      4.22    0.276
     ## # … with 453 more rows
 
-#### 3\. Get regression fit summaries
+#### 3. Get regression fit summaries
 
 Get all the scalar summaries of a regression fit included in
 `summary(score_model)` along with the mean-squared error and root
@@ -118,7 +118,7 @@ get_regression_summaries(score_model)
     ##       <dbl>         <dbl> <dbl> <dbl> <dbl>     <dbl>   <dbl> <dbl> <dbl>
     ## 1     0.011         0.009 0.292 0.540 0.541      5.34   0.021     1   463
 
-#### 4\. Plot parallel slopes models
+#### 4. Plot parallel slopes models
 
 Plot parallel slopes regression models involving one categorical and one
 numerical explanatory/predictor variable (something you cannot do using
@@ -135,7 +135,7 @@ ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
 
 ## Other features
 
-#### 1\. Print markdown friendly tables
+#### 1. Print markdown friendly tables
 
 Want to output cleanly formatted tables in an R Markdown document? Just
 add `print = TRUE` to any of the three `get_regression_*()` functions.
@@ -145,19 +145,18 @@ get_regression_table(score_model, print = TRUE)
 ```
 
 | term      | estimate | std\_error | statistic | p\_value | lower\_ci | upper\_ci |
-| :-------- | -------: | ---------: | --------: | -------: | --------: | --------: |
+|:----------|---------:|-----------:|----------:|---------:|----------:|----------:|
 | intercept |    4.462 |      0.127 |    35.195 |    0.000 |     4.213 |     4.711 |
-| age       |  \-0.006 |      0.003 |   \-2.311 |    0.021 |   \-0.011 |   \-0.001 |
+| age       |   -0.006 |      0.003 |    -2.311 |    0.021 |    -0.011 |    -0.001 |
 
-#### 2\. Predictions on new data
[…]
