+++ DESCRIPTION
@@ -35,7 +35,6 @@ Imports:
     glue
 RoxygenNote: 7.1.1
 Suggests: testthat,
-    covr,
     rmarkdown,
     vdiffr,
     ggplot2movies,
