### 2021-06-04 moderndive/moderndive#110 @lwjohnst86
Remove examples from dataset documentations
Part of the JOSE review https://github.com/openjournals/jose-reviews/issues/115

The examples in the roxygen docs for the datasets aren't really necessary and generally increase the build time of the website/package checks without adding much value. It also isn't a common part of package docs anyway. I'd suggest removing them. The added benefit of removing them is that the ragg package dependency in #105 can be removed (which is needed to generate those documentation pages).
