### 2021-06-04 moderndive/moderndive#111 @lwjohnst86
Remove use of patchwork from vignette
Part of the JOSE review https://github.com/openjournals/jose-reviews/issues/115

Related to #105, in order to automatically build the website, the patchwork package is required for building the vignette. Since it is only needed for that, I'd suggest removing its use so you can remove the `Suggests` dependency for patchwork.
