### 2021-06-04 moderndive/moderndive#112 @lwjohnst86
Strongly encourage using "fitted" or "predicted" compared to "*_hat"
Part of the JOSE review https://github.com/openjournals/jose-reviews/issues/115

This was partly brought up by @lisamr in her review. The language for `y_hat` is very much linked to the math of statistics. However, considering this is a package that could have (potential) broader usage outside of the ModernDive book, most people new to doing statistics will have no idea what that means. There's a reason `broom::augment()` uses `.fitted` instead of `y_hat` or something similar. I'd suggest renaming the column from `get_regression_points()` to be `*_fitted` or `*_predicted` inline with general better communication to non-statisticians, to make the package more learner friendly, and to increase adoption/use of this package. As personal experience, I never understood what y hat was until I learned about it being the "predicted" value based on the regression line. y hat was too related to the equation and I couldn't relate it to the real world.
