### 2021-06-04 moderndive/moderndive#113 @lwjohnst86
(infrastructure, optional comment) Consider moving dataset documentation into one file
Part of JOSE review https://github.com/openjournals/jose-reviews/issues/115

This is more for an infrastructural/developer point of view, but it is a bit confusing having the dataset documentation all over the place. There is the `datasets.R` file which intuitively means that the datasets are documented there, but then there are `pennies.R` etc. This is entirely optional and organizational, but maybe consider moving the dataset documentation into one file rather than multiple.
