### 2021-06-04 moderndive/moderndive#114 @lwjohnst86
(Infrastructural, optional) Consider using markdown to write Roxygen docs
Part of JOSE review https://github.com/openjournals/jose-reviews/issues/115

This is optional and largely infrastructure/developer focused. Right now docs are written using the Rd format of writing. However, if you switch to using Markdown for writing the docs, it makes it easier for e.g. pkgdown to link functions to other help sites. For instance, on the website, the `get_regression_points()` reference page has a "See also" section, with `augment` listed but no link to what `augment` is. If written in Markdown (starting with `use_roxygen_md()`), you get pkgdown to link to augment's help doc with `[broom::augment()]`.
