### 2021-06-04 moderndive/moderndive#115 @lwjohnst86
Larger comments on paper.md
Part of JOSE review: https://github.com/openjournals/jose-reviews/issues/115

Overall, really nice! I like the reference to John Oliver! :stuck_out_tongue: Here are some broader comments on it.

- Similar to @lisamr comments, the paper is a bit too long and much of the "meat" could be linked back to the website.
- One of the biggest feedback is to limit duplication of content between README, vignette, and paper. Keep the content separate.
    - For instance, I see you have `_build.sh`, which builds the paper from the `vignettes/why-moderndive.Rmd`. There's no need for duplication. If most of the information is available in the vignette, link back to the vignette from the paper. I'd keep the paper as minimal as possible and refer to the vignette mostly.
    - There are small differences that should be found between the vignette and the paper (e.g. Statement of Need and Contributions sections in paper but not vignette).
    - You don't really need code output in the paper since it is already in the vignette.
    - By splitting out the vignette text from the paper text, then you can delete the `vignette/README.md` and the `vignette/_build.sh` files.
- I wouldn't recommend saving the JOSE pdf in your repo, since it is already available on GitHub anyway.

A big plus is with emphasizing CI and reducing the p-value interpretation! Very nice :grin: (I personally would remove the p-value entirely, but that's a different debate).
