### 2021-06-04 6d05ed0
make it so code chunk is only for display in vignette
