### 2021-06-04 028a30e
removed suggest dependencies that were only used for data processing or in vignette. Viridis is part of ggplot2
