### 2021-06-04 moderndive/moderndive#117 @lwjohnst86
Repo needs a CONTRIBUTING guidelines file
Part of JOSE review https://github.com/openjournals/jose-reviews/issues/115

There are no guidelines for now to contribute. `use_tidy_contributing()` gives a good template to start from.
