### 2021-06-04 moderndive/moderndive#118 @lwjohnst86
Add one or two sentences on purpose and reason for package being created in README
Part of JOSE review: https://github.com/openjournals/jose-reviews/issues/115

A brief description in the first paragraph in the README file would help with giving users an idea of what the package is used for.
