### 2021-06-04 #115 @rudeboybert
Wow @lwjohnst86 thanks for the very thorough and thoughtful reviews, and double thanks for communicating them as issues & PRs directly in the moderndive package repo!

I hope to address @lisamr's outstanding comments and Luke's PR's issues next week. I'll tag you both as well as @moorepants when I'm done. Many thanks!
