### 2021-06-07 #115 @moorepants
@lisamr and @lwjohnst86 Thanks for your timely and thorough reviews!

@rudeboybert sounds like you are on top of things, looking forward to your responses and adjustments from the feedback.

As was noted by both reviewers, please adjust the paper to fit the JOSE guidelines more closely. It should not be a repeat of the documentation. All other comments look good to me.
