+++ DESCRIPTION
@@ -44,6 +44,5 @@ Suggests:
     patchwork,
     viridis,
     readr,
-    ragg,
-    patchwork
+    ragg
 VignetteBuilder: knitr
