+++ vignettes/paper.md
@@ -1,7 +1,7 @@
 ---
 title: "Take a `moderndive` into introductory linear regression with R"
 author: "Albert Y. Kim, Chester Ismay, and Max Kuhn"
-date: "2021-05-14"
+date: "2021-06-17"
 vignette: >
   %\VignetteIndexEntry{Take a `moderndive` into introductory linear regression with R}
   %\VignetteEncoding{UTF-8}
@@ -193,7 +193,7 @@ ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
 
 \begin{figure}
 
-{\centering \includegraphics[width=1\linewidth]{Figures/interaction-model-1} 
+{\centering \includegraphics[width=1\linewidth]{paper_files/figure-latex/interaction-model-1} 
 
 }
 
@@ -202,7 +202,7 @@ ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
 
 However, many introductory statistics courses start with the easier to teach "common slope, different intercepts" regression model, also known as the *parallel slopes* model. However, no argument to plot such models exists within `geom_smooth()`.
 
-[Evgeni Chasnovski](https://github.com/echasnovski){target="_blank"} thus wrote a custom `geom_` extension to `ggplot2` called `geom_parallel_slopes()`; this extension is included in the `moderndive` package. Much like `geom_smooth()` from the `ggplot2` package, you add `geom_parallel_slopes()` as another layer to the code, resulting in \autoref{fig:parallel-slopes-model}.
+[Evgeni Chasnovski](https://github.com/echasnovski){target="_blank"} thus wrote a custom `geom_` extension to `ggplot2` called `geom_parallel_slopes()`; this extension is included in the `moderndive` package. Much like `geom_smooth()` from the `ggplot2` package, you add `geom_parallel_slopes()` as a layer to the code, resulting in \autoref{fig:parallel-slopes-model}.
 
 
 ```r
@@ -215,7 +215,7 @@ ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
 
 \begin{figure}
 
-{\centering \includegraphics[width=1\linewidth]{Figures/parallel-slopes-model-1} 
+{\centering \includegraphics[width=1\linewidth]{paper_files/figure-latex/parallel-slopes-model-1} 
 
 }
 
@@ -394,7 +394,7 @@ ggplot(score_model_points, aes(x = residual)) +
 
 \begin{figure}
 
-{\centering \includegraphics[width=1\linewidth]{Figures/residuals-1-1} 
+{\centering \includegraphics[width=1\linewidth]{paper_files/figure-latex/residuals-1-1} 
 
 }
 
@@ -413,13 +413,12 @@ ggplot(score_model_points, aes(x = age, y = residual)) +
 
 \begin{figure}
 
-{\centering \includegraphics[width=1\linewidth]{Figures/residuals-2-1} 
+{\centering \includegraphics[width=1\linewidth]{paper_files/figure-latex/residuals-2-1} 
 
 }
 
 \caption{Partial residual residual plot over age.}\label{fig:residuals-2}
 \end{figure}
-
 
 ## 4. A quick-and-easy Kaggle predictive modeling competition submission!
 
@@ -430,7 +429,6 @@ The fourth common student question:
 With the fields of machine learning and artificial intelligence gaining prominence, the importance of predictive modeling cannot be understated. Therefore, we've designed the `get_regression_points()` function to allow for a `newdata` argument to quickly apply a previously fitted model to new observations. 
 
[…]
