+++ vignettes/_build.sh
@@ -3,4 +3,4 @@
 # this file executable:
 # chmod +x ./_build.sh
 
-Rscript -e 'rmarkdown::render(input = "why-moderndive.Rmd", output_file = "paper")'
+Rscript -e 'rmarkdown::render(input = "moderndive.Rmd", output_file = "paper")'
