+++ man/pennies.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/pennies.R
+% Please edit documentation in R/datasets.R
 \docType{data}
 \name{pennies}
 \alias{pennies}
