+++ man/tactile_prop_red.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/bowl.R
+% Please edit documentation in R/datasets.R
 \docType{data}
 \name{tactile_prop_red}
 \alias{tactile_prop_red}
