+++ man/bowl_sample_1.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/bowl.R
+% Please edit documentation in R/datasets.R
 \docType{data}
 \name{bowl_sample_1}
 \alias{bowl_sample_1}
