+++ R/datasets.R
@@ -11,11 +11,6 @@
 #' }
 #' @note The original \code{pennies_sample} has been renamed \code{\link{orig_pennies_sample}}
 #' as of \code{moderndive} v0.3.0.
-#' @examples
-#' library(ggplot2)
-#'
-#' ggplot(pennies_sample, aes(x = year)) +
-#'   geom_histogram(binwidth = 5, boundary = 2000)
 "pennies_sample"
 
 
@@ -35,16 +30,6 @@
 #'   \item{year}{Year on resampled penny}
 #' }
 #' @seealso \code{\link{pennies_sample}}
-#' @examples
-#' library(ggplot2)
-#' library(dplyr)
-#' bootstrap_sample_means <- pennies_resamples %>%
-#'   group_by(name) %>%
-#'   summarize(sample_mean = mean(year))
-#'
-#' ggplot(bootstrap_sample_means, aes(x = sample_mean)) +
-#'   geom_histogram(binwidth = 2.5) +
-#'   labs(x = "sample mean year", title = "Bootstrap distribution of sample mean year")
 "pennies_resamples"
 
 
@@ -60,27 +45,6 @@
 #'   \item{age_in_2011}{Age in 2011}
 #' }
 #' @source StatCrunch \url{https://www.statcrunch.com/app/index.php?dataid=301596}
-#' @examples
-#' library(dplyr)
-#' library(ggplot2)
-#'
-#' # Take 25 different samples of size n = 50 pennies from population
-#' many_samples <- pennies %>%
-#'   rep_sample_n(size = 50, reps = 25)
-#' many_samples
-#'
-#' # Compute mean year of minting for each sample
-#' sample_means <- many_samples %>%
-#'   group_by(replicate) %>%
-#'   summarize(mean_year = mean(year))
-#'
-#' # Plot sampling distribution
-#' ggplot(sample_means, aes(x = mean_year)) +
-#'   geom_histogram(binwidth = 1, color = "white") +
-#'   labs(
-#'     x = expression(bar(x)), y = "Number of samples",
-#'     title = "Sampling distribution of x_bar based 25 samples of size n = 50"
-#'   )
 "pennies"
 
 
@@ -97,27 +61,6 @@
 #' }
 #' @source StatCrunch \url{https://www.statcrunch.com/app/index.php?dataid=301596}
 #' @seealso \code{\link{pennies}}
-#' @examples
-#' library(dplyr)
-#' library(ggplot2)
-#'
-#' # Take 50 different resamples/bootstraps from the original sample
-#' many_bootstraps <- orig_pennies_sample %>%
-#'   rep_sample_n(size = 40, replace = TRUE, reps = 50)
-#' many_bootstraps
-#'
-#' # Compute mean year of minting for each bootstrap sample
-#' bootstrap_means <- many_bootstraps %>%
-#'   group_by(replicate) %>%
-#'   summarize(mean_year = mean(year))
-#'
-#' # Plot sampling distribution
-#' ggplot(bootstrap_means, aes(x = mean_year)) +
-#'   geom_histogram(binwidth = 1, color = "white") +
-#'   labs(
-#'     x = expression(bar(x)), y = "Number of samples",
-#'     title = "Bootstrap distribution of x_bar based 50 resamples of size n = 40"
-#'   )
 "orig_pennies_sample"
 
 
@@ -137,29 +80,6 @@
 #'   marked on the balls themselves}
 #'   \item{color}{color of ball: red or white}
 #' }
-#' @examples
-#' library(dplyr)
-#' library(ggplot2)
-#'
-#' # Take 10 different samples of size n = 50 balls from bowl
-#' bowl_samples_simulated <- bowl %>%
-#'   rep_sample_n(50, reps = 10)
-#'
[…]
