+++ man/house_prices.Rd
@@ -42,16 +42,4 @@ This dataset contains house sale prices for King County, which includes
 Seattle. It includes homes sold between May 2014 and May 2015. This dataset
 was obtained from Kaggle.com \url{https://www.kaggle.com/harlfoxem/housesalesprediction/data}
 }
-\examples{
-library(dplyr)
-library(ggplot2)
-
-# Create variable log of house price
-house_prices <- house_prices \%>\%
-  mutate(log_price = log(price))
-
-# Plot histogram of log of house price
-ggplot(house_prices, aes(x = log_price)) +
-  geom_histogram()
-}
 \keyword{datasets}

[…]
