### 2021-06-17 moderndive/moderndive#111 @rudeboybert
Thanks for the suggestion. However we really want [Figures 7 & 8](https://raw.githubusercontent.com/openjournals/jose-papers/jose.00115/jose.00115/10.21105.jose.00115.pdf) to be presented side-by-side so that the reader can make quick comparisons.
