### 2021-06-17 moderndive/moderndive#112 @rudeboybert
Thanks for the suggestions @lwjohnst86 & @lisamr. However b/c the ModernDive book (in particular the print edition) explains these values as "hat", it would soft break the book of we changed this. So for practical reasons, we're going to keep as is. But philosophically and pedagogically speaking, I agree this is suboptimal.
