### 2021-06-17 9d7496b
Update NEWS
