+++ README.Rmd
@@ -15,7 +15,9 @@ knitr::opts_chunk$set(echo = TRUE, fig.path = "man/figures/")
 [![CRAN RStudio mirror downloads](http://cranlogs.r-pkg.org/badges/moderndive)](http://www.r-pkg.org/pkg/moderndive)
 
 
-An R package of datasets and wrapper functions for [tidyverse](https://www.tidyverse.org/)-friendly introductory linear regression used in "Statistical Inference via Data Science: A ModernDive into R and the Tidyverse" available at [ModernDive.com](https://moderndive.com/).
+## Overview
+
+The [`moderndive`](https://moderndive.github.io/moderndive/) R package consists of datasets and functions for [tidyverse](https://www.tidyverse.org/)-friendly introductory linear regression. These tools leverage the well-developed `tidyverse` and `broom` packages to facilitate 1) working with regression tables that include confidence intervals, 2) accessing regression outputs on an observation level (e.g. fitted/predicted values and residuals), 3) inspecting scalar summaries of regression fit (e.g. R-squared, R-squared adjusted, and mean squared error), and 4) visualizing parallel slopes regression models using `ggplot2`-like syntax. This R package is designed to supplement the book "Statistical Inference via Data Science: A ModernDive into R and the Tidyverse" available at [ModernDive.com](https://moderndive.com/).
 
 
 ## Installation
