### 2021-06-17 moderndive/moderndive#110 @rudeboybert
Great advice. GitHub Actions checks are much quicker now
