+++ R/datasets.R
@@ -2,15 +2,15 @@
 #'
 #' A sample of 50 pennies contained in a 50 cent roll from Florence Bank on
 #' Friday February 1, 2019 in downtown Northampton, Massachusetts, USA
-#' \url{https://goo.gl/maps/AF88fpvVfm12}.
+#' <https://goo.gl/maps/AF88fpvVfm12>.
 #'
 #' @format A data frame of 50 rows representing 50 sampled pennies and 2 variables
 #' \describe{
 #'   \item{ID}{Variable used to uniquely identify each penny.}
 #'   \item{year}{Year of minting.}
 #' }
-#' @note The original \code{pennies_sample} has been renamed \code{\link{orig_pennies_sample}}
-#' as of \code{moderndive} v0.3.0.
+#' @note The original `pennies_sample` has been renamed [orig_pennies_sample()]
+#' as of `moderndive` v0.3.0.
 "pennies_sample"
 
 
@@ -19,8 +19,8 @@
 #'
 #' 35 bootstrap resamples with replacement of sample of 50 pennies contained in
 #' a 50 cent roll from Florence Bank on Friday February 1, 2019 in downtown Northampton,
-#' Massachusetts, USA \url{https://goo.gl/maps/AF88fpvVfm12}. The original sample
-#' of 50 pennies is available in \code{\link{pennies_sample}} .
+#' Massachusetts, USA <https://goo.gl/maps/AF88fpvVfm12>. The original sample
+#' of 50 pennies is available in [pennies_sample()] .
 #'
 #' @format A data frame of 1750 rows representing 35 students' bootstrap
 #' resamples of size 50 and 3 variables
@@ -29,7 +29,7 @@
 #'   \item{name}{Name of student}
 #'   \item{year}{Year on resampled penny}
 #' }
-#' @seealso \code{\link{pennies_sample}}
+#' @seealso [pennies_sample()]
 "pennies_resamples"
 
 
@@ -44,23 +44,23 @@
 #'   \item{year}{Year of minting}
 #'   \item{age_in_2011}{Age in 2011}
 #' }
-#' @source StatCrunch \url{https://www.statcrunch.com/app/index.php?dataid=301596}
+#' @source StatCrunch <https://www.statcrunch.com/app/index.php?dataid=301596>
 "pennies"
 
 
 
-#' A random sample of 40 pennies sampled from the \code{pennies} data frame
+#' A random sample of 40 pennies sampled from the `pennies` data frame
 #'
-#' A dataset of 40 pennies to be treated as a random sample with \code{\link{pennies}} acting
+#' A dataset of 40 pennies to be treated as a random sample with [pennies()] acting
 #' as the population. Data on these pennies were recorded in 2011.
 #'
-#' @format A data frame of 40 rows representing 40 randomly sampled pennies from \code{\link{pennies}} and 2 variables
+#' @format A data frame of 40 rows representing 40 randomly sampled pennies from [pennies()] and 2 variables
 #' \describe{
 #'   \item{year}{Year of minting}
 #'   \item{age_in_2011}{Age in 2011}
 #' }
-#' @source StatCrunch \url{https://www.statcrunch.com/app/index.php?dataid=301596}
-#' @seealso \code{\link{pennies}}
+#' @source StatCrunch <https://www.statcrunch.com/app/index.php?dataid=301596>
+#' @seealso [pennies()]
 "orig_pennies_sample"
 
 
@@ -71,7 +71,7 @@
 #' A sampling bowl of red and white balls
 #'
 #' A sampling bowl used as the population in a simulated sampling exercise. Also
-#' known as the urn sampling framework \url{https://en.wikipedia.org/wiki/Urn_problem}.
+#' known as the urn sampling framework <https://en.wikipedia.org/wiki/Urn_problem>.
 #'
[…]
