### 2021-06-17 moderndive/moderndive#115 @rudeboybert
Remaining TODOs:
- [ ] Review checklist:
    - [ ] A statement of need: Do the authors clearly state the need for this software and who the target audience is?
    - [ ] References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)
- [ ] https://github.com/openjournals/jose-reviews/issues/115#issuecomment-841377884
