+++ vignettes/paper.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "Take a `moderndive` into introductory linear regression with R"
 author: "Albert Y. Kim, Chester Ismay, and Max Kuhn"
-date: "2021-06-27"
+date: "`r Sys.Date()`"
 vignette: >
   %\VignetteIndexEntry{Take a `moderndive` into introductory linear regression with R}
   %\VignetteEncoding{UTF-8}
@@ -39,7 +39,50 @@ output:
     number_sections: yes
 ---
 
+```{r, include = FALSE}
+# knitr settings
+knitr::opts_chunk$set(
+  # Code output:
+  warning = FALSE,
+  message = FALSE,
+  echo = TRUE,
+  # Figure:
+  out.width = "100%",
+  fig.width = 16 / 2.5,
+  fig.height = 9 / 2.5,
+  fig.align = "center",
+  fig.show = "hold",
+  # Etc:
+  collapse = TRUE,
+  comment = "##"
+  # tidy = FALSE
+)
 
+# Needed packages in vignette
+library(moderndive)
+library(ggplot2)
+library(dplyr)
+library(knitr)
+library(broom)
+
+# Needed packages internally
+library(patchwork)
+
+# Random number generator seed value
+set.seed(76)
+
+# Set ggplot defaults for rticles output:
+if (!knitr::is_html_output()) {
+  # Grey theme:
+  theme_set(theme_light())
+
+  scale_colour_discrete <- ggplot2::scale_colour_viridis_d
+}
+
+
+# Set output width for rticles:
+options(width = 70)
+```
 
 
 # Summary
@@ -56,8 +99,7 @@ Linear regression has long been a staple of introductory statistics courses. Whi
 
 Let's load all the R packages we are going to need.
 
-
-```r
+```{r}
 library(moderndive)
 library(ggplot2)
 library(dplyr)
@@ -67,22 +109,22 @@ library(broom)
 
 Let's consider data gathered from end of semester student evaluations for a sample of 463 courses taught by 94 professors from the University of Texas at Austin [@diez2015openintro]. This data is included in the `evals` data frame from the `moderndive` package.
 
+```{r, echo=FALSE}
+evals_sample <- evals %>%
+  select(ID, prof_ID, score, age, bty_avg, gender, ethnicity, language, rank) %>%
+  sample_n(5)
+```
 
-
-In the following table, we present a subset of 9 of the 14 variables included for a random sample of 5 courses^[For details on the remaining 5 variables, see the help file by running `?evals`.]:
+In the following table, we present a subset of `r ncol(evals_sample)` of the `r ncol(evals)` variables included for a random sample of `r nrow(evals_sample)` courses^[For details on the remaining `r ncol(evals) - ncol(evals_sample)` variables, see the help file by running `?evals`.]:
 
 1. `ID` uniquely identifies the course whereas `prof_ID` identifies the professor who taught this course. This distinction is important since many professors taught more than one course.
 1. `score` is the outcome variable of interest: average professor evaluation score out of 5 as given by the students in this course.
 1. The remaining variables are demographic variables describing that course's instructor, including `bty_avg` (average "beauty" score) for that professor as given by a panel of 6 students.^[Note that `gender` was collected as a binary variable at the time of the study (2005).]
 
-
-|  ID| prof_ID| score| age| bty_avg|gender |ethnicity    |language |rank    |
-|---:|-------:|-----:|---:|-------:|:------|:------------|:--------|:-------|
-| 129|      23|   3.7|  62|   3.000|male   |not minority |english  |tenured |
-| 109|      19|   4.7|  46|   4.333|female |not minority |english  |tenured |
-|  28|       6|   4.8|  62|   5.500|male   |not minority |english  |tenured |
[…]
