### 2021-07-10 06f2f7a
First pass at larger comment edits
