+++ vignettes/paper.Rmd
@@ -142,18 +142,10 @@ Let's now study the output of the fitted model `score_model` "the good old-fashi
 summary(score_model)
 ```
 
-Here are 5 common student questions we've heard over the years in our introductory statistics courses based on this output: 
-
-1. "Wow! Look at those p-value stars! Stars are good, so I should try to get many stars, right?"
-1. "How do we extract the values in the regression table?"
-1. "Where are the fitted/predicted values and residuals?"
-1. "How do I apply this model to a new set of data to make predictions?"
-1. "What is all this other stuff at the bottom?"
-
 
 ## Regression analysis using `moderndive`
 
-To address these questions, we've included three functions in the `moderndive` package that take a fitted model object as input and return the same information as `summary.lm()`, but output them in tidyverse-friendly format [@tidyverse2019]. As we'll see later, while these three functions are thin wrappers to existing functions in the `broom` package for converting statistical objects into tidy tibbles, we modified them with the introductory statistics student in mind [@R-broom].
+As an improvement to base R's regression functions, we've included three functions in the `moderndive` package that take a fitted model object as input and return the same information as `summary.lm()`, but output them in tidyverse-friendly format [@tidyverse2019]. As we'll see later, while these three functions are thin wrappers to existing functions in the `broom` package for converting statistical objects into tidy tibbles, we modified them with the introductory statistics student in mind [@R-broom].
 
 1. Get a tidy regression table **with confidence intervals**:
     ```{r}
@@ -167,9 +159,6 @@ get_regression_points(score_model)
     ```{r}
 get_regression_summaries(score_model)
     ```
-
-
-## Bonus: Visualizing parallel slopes models with `moderndive`
 
 Furthermore, say you would like to create a visualization of the relationship between two numerical variables and a third categorical variable with $k$ levels. Let's create this using a colored scatterplot via the `ggplot2` package for data visualization [@R-ggplot2]. Using `geom_smooth(method = "lm", se = FALSE)` yields a visualization of an *interaction model* where each of the $k$ regression lines has their own intercept and slope. For example in \autoref{fig:interaction-model}, we extend our previous regression model by now mapping the categorical variable `ethnicity` to the `color` aesthetic.
 
@@ -193,348 +182,34 @@ ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
   labs(x = "Age", y = "Teaching score", color = "Ethnicity")
 ```
 
-At this point however, students will inevitably ask a sixth question: "When would you ever use a parallel slopes model?" 
 
 
-## Why should you use the `moderndive` package?
 
-To recap this introduction, we believe that the following functions included in the `moderndive` package
+# Repository README
 
-1. `get_regression_table()`
-1. `get_regression_points()`
-1. `get_regression_summaries()`
-1. `geom_parallel_slopes()`
+In the GitHub repository README, we present an in depth discussion of six features of the `moderndive` package:
 
-are effective pedagogical tools that can help address the six common questions posed by students about introductory linear regression performed in R. We now argue why.
+1. Focus less on p-value stars, more confidence intervals
+2. Outputs as tibbles
+3. Produce residual analysis plots from scratch using `ggplot2`
+4. A quick-and-easy Kaggle predictive modeling competition submission!
+5. Visual model selection: plot parallel slopes & interaction regression models
+6. Produce metrics on the quality of regression model fits
 
[…]
