+++ vignettes/paper.md
@@ -1,7 +1,7 @@
 ---
 title: "Take a `moderndive` into introductory linear regression with R"
 author: "Albert Y. Kim, Chester Ismay, and Max Kuhn"
-date: "2021-06-27"
+date: "2021-07-09"
 vignette: >
   %\VignetteIndexEntry{Take a `moderndive` into introductory linear regression with R}
   %\VignetteEncoding{UTF-8}
@@ -120,18 +120,10 @@ summary(score_model)
 ## F-statistic: 5.342 on 1 and 461 DF,  p-value: 0.02125
 ```
 
-Here are 5 common student questions we've heard over the years in our introductory statistics courses based on this output: 
-
-1. "Wow! Look at those p-value stars! Stars are good, so I should try to get many stars, right?"
-1. "How do we extract the values in the regression table?"
-1. "Where are the fitted/predicted values and residuals?"
-1. "How do I apply this model to a new set of data to make predictions?"
-1. "What is all this other stuff at the bottom?"
-
 
 ## Regression analysis using `moderndive`
 
-To address these questions, we've included three functions in the `moderndive` package that take a fitted model object as input and return the same information as `summary.lm()`, but output them in tidyverse-friendly format [@tidyverse2019]. As we'll see later, while these three functions are thin wrappers to existing functions in the `broom` package for converting statistical objects into tidy tibbles, we modified them with the introductory statistics student in mind [@R-broom].
+As an improvement to base R's regression functions, we've included three functions in the `moderndive` package that take a fitted model object as input and return the same information as `summary.lm()`, but output them in tidyverse-friendly format [@tidyverse2019]. As we'll see later, while these three functions are thin wrappers to existing functions in the `broom` package for converting statistical objects into tidy tibbles, we modified them with the introductory statistics student in mind [@R-broom].
 
 1. Get a tidy regression table **with confidence intervals**:
     
@@ -173,9 +165,6 @@ To address these questions, we've included three functions in the `moderndive` p
     ## # ... with 1 more variable: nobs <dbl>
     ```
 
-
-## Bonus: Visualizing parallel slopes models with `moderndive`
-
 Furthermore, say you would like to create a visualization of the relationship between two numerical variables and a third categorical variable with $k$ levels. Let's create this using a colored scatterplot via the `ggplot2` package for data visualization [@R-ggplot2]. Using `geom_smooth(method = "lm", se = FALSE)` yields a visualization of an *interaction model* where each of the $k$ regression lines has their own intercept and slope. For example in \autoref{fig:interaction-model}, we extend our previous regression model by now mapping the categorical variable `ethnicity` to the `color` aesthetic.
 
 
@@ -189,7 +178,7 @@ ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
 
 \begin{figure}
 
-{\centering \includegraphics[width=1\linewidth]{/Users/rudeboybert/Documents/moderndive/moderndive/vignettes/paper_files/figure-latex/interaction-model-1} 
+{\centering \includegraphics[width=1\linewidth]{paper_files/figure-latex/interaction-model-1} 
 
 }
 
@@ -211,507 +200,41 @@ ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
 
 \begin{figure}
 
-{\centering \includegraphics[width=1\linewidth]{/Users/rudeboybert/Documents/moderndive/moderndive/vignettes/paper_files/figure-latex/parallel-slopes-model-1} 
+{\centering \includegraphics[width=1\linewidth]{paper_files/figure-latex/parallel-slopes-model-1} 
 
 }
 
[…]
