+++ README.Rmd
@@ -139,10 +139,6 @@ evals_sample %>%
 
 ### 1. Focus less on p-value stars, more confidence intervals
 
-The first common student question:
-
-> "Wow! Look at those p-value stars! Stars are good, so I should try to get many stars, right?"
-
 We argue that the `summary.lm()` output is deficient in an introductory statistics setting because:
 
 1. The `Signif. codes:  0 '' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1` only encourage **p-hacking**. In case you have not yet been convinced of the perniciousness of p-hacking, perhaps comedian [John Oliver can convince you](https://www.youtube.com/watch?v=0Rnq1NpHdmw){target="_blank"}.  
@@ -162,10 +158,6 @@ get_regression_table(score_model, conf.level = 0.99)
 
 
 ### 2. Outputs as tibbles
-
-The second common student question:
-
-> "How do we extract the values in the regression table?"
 
 While one might argue that extracting the intercept and slope coefficients can be "simply" done using `coefficients(score_model)`, what about the standard errors? For example, a Google query of "_how do I extract standard errors from lm in R_" yielded results from [the R mailing list](https://stat.ethz.ch/pipermail/r-help/2008-April/160538.html){target="_blank"} and from [Cross Validated](https://stats.stackexchange.com/questions/27511/extract-standard-errors-of-coefficient-linear-regression-r){target="_blank"} suggesting we run:
 
@@ -196,10 +188,6 @@ get_regression_table(score_model) %>%
 
 
 ### 3. Produce residual analysis plots from scratch using `ggplot2`
-
-The third common student question:
-
-> "Where are the fitted/predicted values and residuals?"
 
 How can we extract point-by-point information from a regression model, such as the fitted/predicted values and the residuals? (Note we only display the first 10 out of 463 of such values for brevity's sake.)
 
@@ -249,10 +237,6 @@ ggplot(score_model_points, aes(x = age, y = residual)) +
 ```
 
 ### 4. A quick-and-easy Kaggle predictive modeling competition submission!
-
-The fourth common student question:
-
-> "How do I apply this model to a new set of data to make predictions?"
 
 With the fields of machine learning and artificial intelligence gaining prominence, the importance of predictive modeling cannot be understated. Therefore, we've designed the `get_regression_points()` function to allow for a `newdata` argument to quickly apply a previously fitted model to new observations. 
 
@@ -304,10 +288,6 @@ knitr::include_graphics("vignettes/leaderboard_orig.png")
 ```
 
 ### 5. Visual model selection: plot parallel slopes & interaction regression models
-
-Finally, the last common student question:
-
-> "When would you ever use a parallel slopes model?"
 
 For example, recall the earlier visualizations of the interaction and parallel slopes models for teaching score as a function of age and ethnicity we saw in Figures \ref{fig:interaction-model} and \ref{fig:parallel-slopes-model}. Let's present both visualizations side-by-side in \autoref{fig:interaction-and-parallel-slopes-model-1}.
 
@@ -419,10 +399,6 @@ While many students will inevitably find these results depressing, in our opinio
 
 ### 6. Produce metrics on the quality of regression model fits
 
-The fifth common student question:
-
-> "What is all this other stuff at the bottom?"
-
 Recall the output of the standard `summary.lm()` from earlier:
 
 ```{r, echo=FALSE}
@@ -475,6 +451,8 @@ The `geom_parallel_slopes()` is a custom built `geom` extension to the `ggplot2`
 ***
 
 
+
+## Contributor code of conduct
 
 
 Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
