### 2021-07-12 c7f4b53
Update vignettes/paper.md
