### 2021-07-12 2cdd852
Update vignettes/paper.Rmd
