### 2021-07-12 8b95ea0
Update vignettes/paper.Rmd
