+++ README.md
@@ -31,8 +31,6 @@ This R package is designed to supplement the book “Statistical Inference
 via Data Science: A ModernDive into R and the Tidyverse” available at
 [ModernDive.com](https://moderndive.com/).
 
------
-
 ## Installation
 
 Get the released version from CRAN:
@@ -48,8 +46,6 @@ Or the development version from GitHub:
 # install.packages("remotes")
 remotes::install_github("moderndive/moderndive")
 ```
-
------
 
 ## Basic usage
 
@@ -118,6 +114,12 @@ score_model <- lm(score ~ age, data = evals)
     
     ![](man/figures/unnamed-chunk-7-1.png)<!-- -->
 
+## Contributor code of conduct
+
+Please note that this project is released with a [Contributor Code of
+Conduct](CONDUCT.md). By participating in this project you agree to
+abide by its terms.
+
 -----
 
 ## Six features
@@ -145,8 +147,8 @@ However we first discuss the model and data background
 
 The data consists of end of semester student evaluations for a sample of
 463 courses taught by 94 professors from the University of Texas at
-Austin (Diez, Barr, and Çetinkaya-Rundel 2015). This data is included in
-the `evals` data frame from the `moderndive` package.
+Austin. This data is included in the `evals` data frame from the
+`moderndive` package.
 
 In the following table, we present a subset of 9 of the 14 variables
 included for a random sample of 5 courses\[1\]:
@@ -158,15 +160,16 @@ included for a random sample of 5 courses\[1\]:
     evaluation score out of 5 as given by the students in this course.
 3.  The remaining variables are demographic variables describing that
     course’s instructor, including `bty_avg` (average “beauty” score)
-    for that professor as given by a panel of 6 students.\[2\]
+    for that professor as given by a panel of 6
+students.\[2\]
 
-|  ID | prof\_ID | score | age | bty\_avg | gender | ethnicity    | language | rank         |
-| --: | -------: | ----: | --: | -------: | :----- | :----------- | :------- | :----------- |
-| 346 |       70 |   4.4 |  43 |    3.500 | male   | not minority | english  | tenure track |
-| 335 |       68 |   2.4 |  60 |    1.667 | male   | not minority | english  | tenured      |
-| 226 |       41 |   4.7 |  35 |    4.833 | male   | not minority | english  | tenured      |
-| 142 |       26 |   4.4 |  58 |    3.833 | male   | not minority | english  | tenured      |
-|  54 |       10 |   5.0 |  47 |    5.500 | male   | not minority | english  | teaching     |
+|  ID | prof\_ID | score | age | bty\_avg | gender | ethnicity    | language    | rank         |
+| --: | -------: | ----: | --: | -------: | :----- | :----------- | :---------- | :----------- |
+| 250 |       48 |   3.4 |  50 |    3.167 | female | not minority | english     | teaching     |
+| 245 |       48 |   4.4 |  50 |    3.167 | female | not minority | english     | teaching     |
+|  10 |        4 |   4.5 |  40 |    3.167 | female | not minority | english     | tenured      |
+| 339 |       70 |   4.5 |  43 |    3.500 | male   | not minority | english     | tenure track |
+|  75 |       14 |   3.5 |  49 |    4.000 | male   | not minority | non-english | tenured      |
 
 ### 1\. Focus less on p-value stars, more confidence intervals
 
@@ -235,9 +238,8 @@ sqrt(diag(vcov(score_model)))
 We argue that this task shouldn’t be this hard, especially in an
 introductory statistics setting. To rectify this, the three
[…]
