+++ vignettes/paper.bib
@@ -24,7 +24,6 @@
 	Author = {{American Statistical Association Undergraduate Guidelines Workgroup}},
 	Institution = {American Statistical Association},
 	Title = {{``Guidelines for Assessment and Instruction in Statistics Education (GAISE) in Statistics Education (GAISE) College Report College Report 2016''}},
-	Url = {https://www.amstat.org/asa/education/Guidelines-for-Assessment-and-Instruction-in-Statistics-Education-Reports.aspx},
 	Year = {2016}
 	}
 
