### 2021-07-12 #115 @rudeboybert
Hello @moorepants @lisamr @lwjohnst86, many thanks for all your work in this review. I believe I've addressed all the outstanding issues, in particular cutting down the manuscript length from ~2500 to 1026 words. Please let me know if I need to do anything else.
