+++ vignettes/paper.Rmd
@@ -2,10 +2,6 @@
 title: "Take a `moderndive` into introductory linear regression with R"
 author: "Albert Y. Kim, Chester Ismay, and Max Kuhn"
 date: "`r Sys.Date()`"
-vignette: >
-  %\VignetteIndexEntry{Take a `moderndive` into introductory linear regression with R}
-  %\VignetteEncoding{UTF-8}
-  %\VignetteEngine{knitr::rmarkdown}
 tags:
   - R
   - Rstats
@@ -33,10 +29,10 @@ affiliations:
  - name: Software Engineer, RStudio, USA.
    index: 3
 bibliography: paper.bib
-output:
-  rticles::joss_article:
-    keep_md: yes
-    number_sections: yes
+# output:
+#   rticles::joss_article:
+#     keep_md: yes
+#     number_sections: yes
 ---
 
 ```{r, include = FALSE}
@@ -213,6 +209,6 @@ Albert Y. Kim and Chester Ismay contributed equally to the development of the `m
 
 # Acknowledgments
 
-Many thanks to Jenny Smetzer [\@smetzer180](https://github.com/smetzer180){target="_blank"}, Luke W. Johnston [\@lwjohnst86](https://github.com/lwjohnst86){target="_blank"}, and Lisa Rosenthal [\@lisamr](https://github.com/lisamr){target="_blank"} for their helpful feedback for this vignette and to Evgeni Chasnovski [\@echasnovski](https://github.com/echasnovski){target="_blank"} for contributing the `geom_parallel_slopes()` function via GitHub [pull request](https://github.com/moderndive/moderndive/pull/55){target="_blank"}. The authors do not have any financial support to disclose.
+Many thanks to Jenny Smetzer [\@smetzer180](https://github.com/smetzer180){target="_blank"}, Luke W. Johnston [\@lwjohnst86](https://github.com/lwjohnst86){target="_blank"}, and Lisa Rosenthal [\@lisamr](https://github.com/lisamr){target="_blank"} for their helpful feedback for this paper and to Evgeni Chasnovski [\@echasnovski](https://github.com/echasnovski){target="_blank"} for contributing the `geom_parallel_slopes()` function via GitHub [pull request](https://github.com/moderndive/moderndive/pull/55){target="_blank"}. The authors do not have any financial support to disclose.
 
 # References
