+++ pkgdown.yml
@@ -3,7 +3,8 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   moderndive: moderndive.html
-last_built: 2021-06-18T02:18Z
+  paper: paper.html
+last_built: 2021-07-13T00:59Z
 urls:
   reference: moderndive.github.io/moderndive//reference
   article: moderndive.github.io/moderndive//articles
