+++ vignettes/paper.bib
@@ -4,7 +4,8 @@
   isbn={9781943450039},
   url={https://books.google.com/books?id=xNMWswEACAAJ},
   year={2015},
-  publisher={OpenIntro, Incorporated}
+  publisher={OpenIntro, Incorporated},
+  doi = {10.5070/T573020084}
 }
 
 @book{ismay2019moderndive,
@@ -15,7 +16,8 @@
   url={https://books.google.com/books?id=SWHGDwAAQBAJ},
   year={2019},
   publisher={CRC Press},
-  url = {https://moderndive.com/}
+  url = {https://moderndive.com/},
+  doi = {10.1080/00224065.2020.1848366}
 }
 
 
@@ -50,7 +52,7 @@
   author = {David Robinson and Alex Hayes},
   year = {2019},
   note = {R package version 0.5.2},
-  url = {https://CRAN.R-project.org/package=broom},
+  url = {https://CRAN.R-project.org/package=broom}
 }
 
 @Article{tidyverse2019,
