### 2021-07-13 493417d
Added doi links to bib
