+++ vignettes/paper.bib
@@ -66,11 +66,12 @@
 }
 
 @Manual{R-ggplot2,
-  title = {ggplot2: Create Elegant Data Visualisations Using the Grammar of Graphics},
+  title = {{ggplot2}: Create Elegant Data Visualisations Using the Grammar of Graphics},
   author = {Hadley Wickham and Winston Chang and Lionel Henry and Thomas Lin Pedersen and Kohske Takahashi and Claus Wilke and Kara Woo and Hiroaki Yutani},
   year = {2019},
   note = {R package version 3.2.1},
   url = {https://CRAN.R-project.org/package=ggplot2},
+  doi = {10.1007/978-0-387-98141-3}
 }
 
 @Manual{R-tibble,
