### 2021-07-13 #115 @rudeboybert
Hi @moorepants, sounds good. I've broken down my response in 4 sections:

1. Responses (indicating changes) to first round of comments by @lisamr on 5/13
1. Responses (indicating changes) to second round of comments by @lisamr on 5/14
1. Responses (indicating changes) to comments by @lwjohnst86 on 6/4
1. Explanations for unchecked boxes in checklist

# 1. Responses to first round of comments by @lisamr on 5/13

> # General checks
> * authorship: It is currently unclear how Max Kuhn contributed, as they are not listed as a contributor on the Github repository and there isn't an author contribution section in the manuscript to clarify.

I added a "Author contributions" section towards the end of the paper.
> # Documentation
> * Statement of need: It's missing in the Readme file and in the MS

A "Statement of need" section has been added at the outset of the manuscript.
> * Installation instructions: I couldn't find a clearly stated list of dependencies in the repository, but maybe I was looking in the wrong place. I checked under the "dependency graph" tab by clicking on the "insight" button, and no dependencies were listed there either. https://github.com/moderndive/moderndive/network/dependencies

All the dependencies are automatically installed when installing the `moderndive` package, the list of all dependencies can be seen in the [`DESCRIPTION`](https://github.com/moderndive/moderndive/blob/master/DESCRIPTION) file
> * Functional documentation and tests: I'm not qualified to assess this since I don't write packages (I'm more of a user than developer). @lwjohnst86 do you want to make sure to cover that component?

@lwjohnst86 checked this checkbox.
>
> ## Minor details on the readme file
> * Under Demo, it's stated
>
> > Among the many useful features of the moderndive package outlined in our essay “Why should you use the moderndive package for intro linear regression?” we highlight three functions in particular as covered there. We also mention the geom_parallel_slopes() function as #4.
>
> Just say there are 4 functions covered. I'm not sure why `geom_parallel_slopes` is singled out.

That was a typo: The README has been updated (in the Basic Usage section) to highlight 4 functions, not 3.
>
> * for the function `get_regression_table()` I suggest at least in the help documentation, it should be stated what the alpha level is. I know it's probably 0.05, but it should be stated, or even allowed to vary since the point of the function is to understand confidence intervals between a given range.

Thank you for this great suggestion. We updated the functionality of this function to allow for a `conf.level` argument in this [commit](https://github.com/moderndive/moderndive/pull/104/commits/12a972f48f9d092475be5cbebf2ac9956c807e96), and updated the README and manuscript accordingly. This is only currently available in the development version of the package `v0.5.1.9000` for now. This will be included in the next `v0.5.2.` version on CRAN.

> * for the function `get_regression_points()`: If this package is aimed at students very new to statistics, it might be helpful to include the following in the readme.md file: "...including fitted/predicted values (e.g. 'score_hat')"

Thank you, similar language was added to the README in this [commit](https://github.com/moderndive/moderndive/pull/104/commits/f0a775d7758f17ba0c538c44ab309485c1f79a6f)
> * for the function `get_regression_summaries()`: excuse me if I'm wrong, but should `df = 2`? The output from `summary(score_model)` shows 1 and 461 DF.

You are correct. This seems to have been a bug in the `broom::glance()` function that has since been fixed; the README has been updated.

# 2. Responses to second round of comments by @lisamr on 5/14

> @moorepants got it, so the "functionality documentation" checklist item is referring to the object documentation. @rudeboybert below is my feedback:
>
> * documentation for various datasets look great--good detail on explaining each column and providing source of the data

Thank you
> * `geom_categorical_model()`: It is not obvious how the solid and dashed lines differ or how they were determined and it's not clearly stated in the documentation. I thought maybe pairwise contrasts were being performed, but when I look at two categories that are not significantly different (e.g. `ggplot(evals, aes(pic_outfit, score )) + geom_point() + geom_categorical_model()`), their lines are still different. Then I thought maybe each category gets its own linetype, but when I look at `ggplot(mpg, aes(x = drv, y = hwy)) + geom_point() + geom_categorical_model()`, again there are only 2 types.

Since this function is not used in the README nor is it used in the manuscript, I think it's beyond the scope of this review and thus this point can be left unaddressed. Please let me know if you disagree.

> * `get_regression_summaries()`: All of the columns are self-explanatory except for df. More details under the "Value" section would be nice.

The following note is now included in the README in the "Produce metrics on the quality of regression model fits":  and the degrees of freedom `df`

> * `get_regression_table()`: Again, either state that the upper and lower bounds represent the 95% confidence interval, or even better, allow the interval ranges to vary, which in my opinion will help break down this (artificial) idea that 95% CI and the associated p-values are all that matter.

Thanks for this great point. This was addressed above.

>
> # Larger structural comments on the paper
> I went through the paper last night and my first impression was that it was long (~2500 words of text) and overlapped quite a lot with the text and code in the readme file. From what I understand, "Detailed documentation should be present in the repository of the submitted software or module, is reviewed there, and does not need to be included in the paper" (link [here](https://openjournals.readthedocs.io/en/jose/review_criteria.html)). It also lacked a Statement of Need section. I thought a lot of the information presented could have been summed up in a shorter summary paragraph. I also thought the code/output could have been condensed by directing the reader to the repository's documentation.
>
> Here are some key descriptions of the paper, quoted from https://openjournals.readthedocs.io/en/jose/submitting.html:
>
> * Write a short Markdown file titled paper.md with title, author names and affiliations, containing a description of the software, statement of need, and key references.
> * JOSE papers contain a limited set of metadata, plus Summary & Reference sections. We explicitly do not publish long-form articles, because the scholarship represented by a JOSE article is contained in the software or learning modules themselves. Expected length is around **1000 words max**.
>
> @moorepants lmk if I'm misunderstanding anything or maybe clarify how the software paper should differ from the readme and other vignettes?

Thanks for this feedback. I have addressed the Statement of Need section in my previous message. Furthermore, the main paper has been cut from ~2500 words to 1026 words with a lot of the content being moved to the repo README

# 3. Responses to comments by @lwjohnst86 on 6/4

> @rudeboybert @labarba @lisamr I finished my review. All my issues and feedback were either created as Issues or PRs in the moderndive package repo. Overall looks great! If the issues and PRs are addressed, I would say it's good to go!

Thanks for filing all your comments and suggestions as GitHub issues and PR's. I will summarize what I did to address all the above referenced issues from 6/4:

- Infrastructure to build website https://github.com/moderndive/moderndive/issues/105: **Implemented and merged**
- Vignette images not showing https://github.com/moderndive/moderndive/issues/106: **Fixed and merged**
- Code coverage was not actually being generated, fix with this action https://github.com/moderndive/moderndive/issues/107: **Fixed and merged**
- Better navigating reference index objects by grouping them (for website) https://github.com/moderndive/moderndive/issues/108: **Fixed and merged**
- Removing 'judgemental' language from text https://github.com/moderndive/moderndive/issues/109: **Removed such language and merged**
- Remove examples from dataset documentations https://github.com/moderndive/moderndive/issues/110: **Removed all example code from dataset documentation. Now package build and check are much quicker. Issue closed**
- Remove use of patchwork from vignette https://github.com/moderndive/moderndive/issues/111: **Thanks for suggestion, but we really want the two plots side-by-side in order to facilitate visual model selection. Closed**
- Strongly encourage using "fitted" or "predicted" compared to "*_hat" https://github.com/moderndive/moderndive/issues/112: **While this is a good idea, it would cause the package to be misaligned with the book, both the print and online edition at moderndive.com. Closed**
- (infrastructure, optional comment) Consider moving dataset documentation into one file https://github.com/moderndive/moderndive/issues/113: **Great idea, done. Close**
- (Infrastructural, optional) Consider using markdown to write Roxygen docs https://github.com/moderndive/moderndive/issues/114: **Another great suggestion which we implemented. Closed**
- Larger comments on paper.md https://github.com/moderndive/moderndive/issues/115. **The main paper has been cut from ~2500 words to 1026 words with a lot of the content being moved to the repo README**
- General tidy up and removing dependencies https://github.com/moderndive/moderndive/issues/116: **Implemented and merged.**
- Repo needs a CONTRIBUTING guidelines file https://github.com/moderndive/moderndive/issues/117: **Added to repo. Closed**
- Add one or two sentences on purpose and reason for package being created in README https://github.com/moderndive/moderndive/issues/118: **Added in Overview section of README. Closed**

# 4. Explanations for unchecked boxes in checklist

## Unchecked boxes for @lisamr

> A statement of need: Do the authors clearly state the need for this software and who the target audience is?

Added to manuscript

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I've addressed all suggestions to clarify points of confusion in help files (where appropriate) or in the README (for more verbose explanations).

## Unchecked boxes for @lwjohnst86

> A statement of need: Do the authors clearly state the need for this software and who the target audience is?

Added to manuscript
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

[Added](https://github.com/moderndive/moderndive/blob/master/CONTRIBUTING.md) to main repo.
> References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

All books and articles have DOI links. DOI links for
