### 2021-07-14 #115 @moorepants
Thank you for the thorough and clear write up and addressing each point brought up by the reviewers.

@lisamr and @lwjohnst86, can you both take a second pass here at the changes made and report on whether this addresses your prior comments? Please check off the relevant boxes in the checklist too, if addressed.

Please indicate how much time you'd like to finalize your review.
