### 2021-07-14 #115 @lisamr
@rudeboybert great job at tending to the comments and cutting down the MS. I have two small remaining comments that if resolved, I think would mean the paper is good to go.

- I still haven't checked the box for Documentation:Statement of Need. My understanding is that you need a Statement of Need in both the README and the software paper. Otherwise, there wouldn't be a checkbox for both (under Documentation and Software paper).
- In regard to my comment on `geom_categorical_model`: It's possible that the checkbox for "Functionality documentation" solely encompasses content in the README and software paper. If that's the case, you're fine, and I've actually just checked the box for that already. That said, for a potential user of this package, the function documentation for `geom_categorical_model` still leaves room for confusion. For the general utility of this package, a short sentence about why some lines are dashed vs. solid would be nice.
