### 2021-07-14 c995e2d
Added statement of need to README
