+++ README.Rmd
@@ -75,6 +75,11 @@ ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
     ```
 
 
+# Statement of Need
+
+Linear regression has long been a staple of introductory statistics courses. While the curricula of introductory statistics courses has much evolved of late, the overall importance of regression remains the same. Furthermore, while the use of the R statistical programming language for statistical analysis is not new, recent developments such as the `tidyverse` suite of packages have made statistical computation with R accessible to a broader audience. We go one step further by leveraging the `tidyverse` and the `broom` packages to make linear regression accessible to students taking an introductory statistics course. Such students are likely to be new to statistical computation with R; we designed `moderndive` with these students in mind. 
+
+
 
 ## Contributor code of conduct
 
@@ -83,7 +88,6 @@ Please note that this project is released with a [Contributor Code of Conduct](C
 
 
 ***
-
 
 
 ## Six features
