### 2021-07-14 d52a4ae
Added note about solid and dashed lines
