### 2021-07-14 #115 @rudeboybert
Hi @lisamr,

1. I just added the Statement of Need to the README: https://github.com/moderndive/moderndive#statement-of-need
2. I added a note to the documentation for `geom_categorical_model()` explaining what the solid and dashed lines are in this [commit](https://github.com/moderndive/moderndive/commit/d52a4aef3b234b950095703af3a02ee3b4102b53)

Many thanks for all your work on this review
