+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   moderndive: moderndive.html
   paper: paper.html
-last_built: 2021-07-13T15:13Z
+last_built: 2021-07-14T23:11Z
 urls:
   reference: moderndive.github.io/moderndive//reference
   article: moderndive.github.io/moderndive//articles
