### 2021-07-15 #115 @lwjohnst86
@rudeboybert I've checked all the boxes though you still have almost identical text ("six features") between your vignette and your README which I think you should de-duplicate, or rather "don't repeat yourself" :wink: Other than that, it looks great!! :grin:
