### 2021-07-16 #115 @rudeboybert
Thanks @lwjohnst86, I will keep the six features text to form a bridge between the README and the manuscript, even at the cost of repeating myself.

@lisamr could you check off the remaining checkbox? After that point I believe @moorepants can move the review to the next phase
