### 2021-07-19 #115 @moorepants
@rudeboybert I've checked over everything, including @lisamr and @lwjohnst86's final sign off and it all looks ready to go!

Next steps:

1. Create a new release for your github repository and report the tag string here.
2. Create a Zenodo archive of the release made in point 1 and report the DOI here.
