+++ README.Rmd
@@ -83,7 +83,7 @@ Linear regression has long been a staple of introductory statistics courses. Whi
 
 ## Contributor code of conduct
 
-Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
+Please note that this project is released with a [Contributor Code of Conduct](CONTRIBUTING.md). By participating in this project you agree to abide by its terms.
 
 
 
