+++ NEWS.md
@@ -1,8 +1,9 @@
-# moderndive 0.5.1.9000
+# moderndive 0.5.2
 
 * Changed printing of non-baseline categorical variable levels in regression table to be cleaner #102
 * Added explicit `conf.level` argument to `get_regression_table()` inherited from `broom::tidy.lm()`
 * Improved main package vignette based on feedback from @lwjohnst86 & @lisamr
+* Added JOSE publication in `vignettes/paper.md`
 * Fixed `pkgdown` and `covr` issues, defragged documentation.
 
 
