+++ README.md
@@ -131,7 +131,7 @@ designed `moderndive` with these students in mind.
 ## Contributor code of conduct
 
 Please note that this project is released with a [Contributor Code of
-Conduct](CONDUCT.md). By participating in this project you agree to
+Conduct](CONTRIBUTING.md). By participating in this project you agree to
 abide by its terms.
 
 -----
@@ -166,15 +166,16 @@ included for a random sample of 5 courses\[1\]:
     evaluation score out of 5 as given by the students in this course.
 3.  The remaining variables are demographic variables describing that
     course’s instructor, including `bty_avg` (average “beauty” score)
-    for that professor as given by a panel of 6 students.\[2\]
+    for that professor as given by a panel of 6
+students.\[2\]
 
-|  ID | prof\_ID | score | age | bty\_avg | gender | ethnicity    | language    | rank     |
-| --: | -------: | ----: | --: | -------: | :----- | :----------- | :---------- | :------- |
-| 337 |       68 |   2.5 |  60 |    1.667 | male   | not minority | english     | tenured  |
-|  78 |       14 |   3.6 |  49 |    4.000 | male   | not minority | non-english | tenured  |
-| 230 |       42 |   3.6 |  39 |    8.167 | female | not minority | english     | teaching |
-| 421 |       85 |   5.0 |  58 |    7.833 | male   | not minority | english     | teaching |
-|  38 |        8 |   4.4 |  51 |    4.000 | female | not minority | english     | tenured  |
+|  ID | prof\_ID | score | age | bty\_avg | gender | ethnicity    | language | rank         |
+| --: | -------: | ----: | --: | -------: | :----- | :----------- | :------- | :----------- |
+| 140 |       25 |   4.8 |  34 |    7.833 | female | not minority | english  | tenure track |
+| 164 |       31 |   4.4 |  63 |    4.333 | male   | not minority | english  | teaching     |
+|  64 |       12 |   4.2 |  37 |    4.333 | male   | not minority | english  | teaching     |
+| 178 |       34 |   4.2 |  47 |    4.333 | female | minority     | english  | tenure track |
+| 440 |       89 |   3.6 |  35 |    7.833 | female | minority     | english  | tenure track |
 
 ### 1\. Focus less on p-value stars, more confidence intervals
 
@@ -282,7 +283,8 @@ get_regression_table(score_model) %>%
 
 How can we extract point-by-point information from a regression model,
 such as the fitted/predicted values and the residuals? (Note we only
-display the first 10 out of 463 of such values for brevity’s sake.)
+display the first 10 out of 463 of such values for brevity’s
+    sake.)
 
 ``` r
 fitted(score_model)
@@ -399,7 +401,8 @@ get_regression_points(score_model, newdata = new_prof)
 Let’s do another example, this time using the Kaggle [House Prices:
 Advanced Regression
 Techniques](https://www.kaggle.com/c/house-prices-advanced-regression-techniques)
-practice competition ( displays the homepage for this competition).
+practice competition ( displays the homepage for this
+competition).
 
 <div class="figure" style="text-align: center">
 
@@ -453,7 +456,8 @@ write_csv(submission, "submission.csv")
 
 After submitting `submission.csv` to the leaderboard for this Kaggle
 competition, we obtain a “root mean squared logarithmic error” (RMSLE)
-score of 0.42918 as seen in .
+score of 0.42918 as seen in
+.
 
 <div class="figure">
 
@@ -461,7 +465,8 @@ score of 0.42918 as seen in .
 
 <p class="caption">
 
-Resulting Kaggle RMSLE score.
+Resulting Kaggle RMSLE
+score.
 
 </p>
 
