### 2021-07-21 06be0d5
Update CRAN comments
