### 2021-07-21 #115 @rudeboybert
Hey @moorepants:
- I just updated the `moderndive` package on [CRAN](https://cran.r-project.org/package=moderndive) to version 0.5.2
- I've tagged the release on GitHub [here](https://github.com/moderndive/moderndive/releases). By tag string I believe you mean [`v0.5.2`](https://github.com/moderndive/moderndive/tags)
- I created a [Zenodo archive](https://zenodo.org/record/5119261) of the release. The DOI is 10.5281/zenodo.5119261
