### 2021-07-21 #115 @moorepants
@labarba This paper is ready to be accepted!

@lisamr and @lwjohnst86 thank you very much for your quality review. We at JOSE value your contribution.

@rudeboybert thank you for your patience with this submission. I apologize for losing track of it originally. The paper and software look great and I hope you are please with the final reviewed result. This is a very nice addition to the literature. Thanks for submitting to JOSE.
