### 2021-07-21 #115 @labarba
hi  @rudeboybert 👋  — Could you edit the metadata of the Zenodo archive so the title and author list matches the JOSE paper?
